Split impl AppState across feature files
Final slice of the src/app/ refactor. Distributes the 1480-line
`impl AppState { … }` block across ten feature-scoped sibling files,
each carrying one `impl AppState` block under the same module:

  completion.rs  LSP completion / hover / definition / references / quickfix / clipboard-ring handlers
  editing.rs     auto-pair, surround, indent, format, comment toggle, shell filter
  files.rs       save / close / session, file watcher, auto-save, memory + version polling
  git.rs         git dialog, hunk navigation, diff peek, branch / gutter refresh
  lsp.rs         LSP lifecycle, diagnostics, rename, code action, semantic tokens, approval
  modal.rs       modal prompts, fuzzy / symbol / lsp pickers, command palette, welcome / changelog / help / settings
  mouse.rs       hit-testing, scroll predicates, screen-to-buffer geometry
  search.rs      in-buffer search and project-wide search-and-replace
  sidebar.rs     sidebar input, paste, refresh
  snippets.rs    snippet expansion + macro replay + jump list

mod.rs now keeps just the `AppState` struct, `AppState::new`, the
`update()` dispatcher, the `App` runner, and a handful of module-level
constants. Down from 6659 lines to 1655. Methods previously bare `fn`
are now `pub(super) fn` so the dispatcher in mod.rs can call them
across module boundaries; methods that were already `pub fn` (called
from src/main.rs) keep that visibility.

No behaviour change — `scripts/check.sh` (fmt + build + clippy + tests
+ UI suite) passes.
