+++ src/app/mod.rs
@@ -11,14 +11,12 @@ use ratatui::layout::Rect;
 
 use crate::{
     clipboard::ClipboardManager,
-    config::{Config, KeymapPreset, Theme, add_to_recent_files, load_recent_files},
+    config::{Config, load_recent_files},
     editor::Editor,
-    editor::viewport::{screen_pos_to_byte_offset, screen_pos_to_line_display_col},
     git::GitGutter,
     input::{
         InputHandler,
         action::{Direction, EditorAction, ScrollDir},
-        keybinding::KeyBindings,
     },
     search::SearchState,
     ui,
@@ -44,15 +42,21 @@ pub const MAX_SIDEBAR_WIDTH: u16 = 80;
 /// the cursor).
 const DOUBLE_CLICK_INTERVAL: Duration = Duration::from_millis(500);
 
+mod completion;
+mod editing;
+mod files;
+mod git;
+mod lsp;
+mod modal;
+mod mouse;
+mod search;
+mod sidebar;
+mod snippets;
 pub mod state;
 mod util;
 
-use state::sidebar::copy_target_path;
 pub use state::*;
-use util::{
-    SCROLL_LINES, completion_kind_label, extract_hover_text, extract_word_at, parse_locations,
-    parse_lsp_position, read_rss_kb, same_file, scroll_action,
-};
+use util::SCROLL_LINES;
 
 // ── AppState ──────────────────────────────────────────────────────────────────
 
@@ -1434,2509 +1438,33 @@ impl AppState {
         }
     }
 
-    /// Run `command` (via `sh -c`) with the current selection on stdin and
-    /// replace the selection with the captured stdout. Single undo entry.
-    fn apply_shell_filter(&mut self, command: &str) {
-        if self.config.disable_shell_filter {
-            self.status_error = Some("Shell filter disabled by config".to_string());
-            return;
-        }
-        let trimmed = command.trim();
-        if trimmed.is_empty() {
-            return;
-        }
-        let buf = &self.editor.active().buffer;
-        let primary = buf.cursors.primary();
-        let range = match primary.selection {
-            Some(s) => s.as_byte_range(),
-            None => return,
-        };
-        if range.is_empty() {
-            return;
-        }
-        let selection_text = {
-            let rope = buf.rope();
-            let cs = rope.byte_to_char(range.start);
-            let ce = rope.byte_to_char(range.end);
-            rope.slice(cs..ce).to_string()
-        };
-
-        use std::io::Write;
-        use std::process::{Command, Stdio};
-        let mut child = match Command::new("sh")
-            .arg("-c")
-            .arg(trimmed)
-            .stdin(Stdio::piped())
-            .stdout(Stdio::piped())
-            .stderr(Stdio::piped())
-            .spawn()
-        {
-            Ok(c) => c,
-            Err(e) => {
-                self.status_error = Some(format!("filter spawn failed: {e}"));
-                return;
-            }
-        };
-        if let Some(mut stdin) = child.stdin.take() {
-            let _ = stdin.write_all(selection_text.as_bytes());
-            // Drop stdin to signal EOF to the child.
-        }
-        let output = match child.wait_with_output() {
-            Ok(o) => o,
-            Err(e) => {
-                self.status_error = Some(format!("filter wait failed: {e}"));
-                return;
-            }
-        };
-        if !output.status.success() {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            let trimmed_err = stderr.trim();
-            self.status_error = Some(if trimmed_err.is_empty() {
-                format!("filter exited {}", output.status)
-            } else {
-                format!("filter: {trimmed_err}")
-            });
-            return;
-        }
-        let new_text = String::from_utf8_lossy(&output.stdout).to_string();
-        let buf = &mut self.editor.active_mut().buffer;
-        buf.begin_batch();
-        buf.move_cursor_to(range.start, false);
-        buf.move_cursor_to(range.end, true);
-        buf.insert_str(&new_text);
-        buf.commit_batch();
-    }
-
     // ── Modal input handling ─────────────────────────────────────────────────
-
-    fn handle_modal_input(&mut self, action: EditorAction) {
-        // Mark prompts auto-complete on the first typed character.
-        if let EditorAction::InsertChar(c) = action {
-            match self.input_mode {
-                InputMode::SetMarkChar => {
-                    self.input_mode = InputMode::Normal;
-                    let handle = self.editor.active();
-                    if let Some(path) = handle.path.clone() {
-                        let off = handle.buffer.cursors.primary().byte_offset;
-                        self.marks.set(&path, c, off);
-                        self.marks.save(&self.workspace);
-                    } else {
-                        self.status_error = Some("Save the file before setting a mark".into());
-                    }
-                    return;
-                }
-                InputMode::JumpToMarkChar => {
-                    self.input_mode = InputMode::Normal;
-                    let active_path = self.editor.active().path.clone();
-                    if let Some(path) = active_path
-                        && let Some(off) = self.marks.get(&path, c)
-                    {
-                        self.push_current_to_jump_list();
-                        let handle = self.editor.active_mut();
-                        let rope = handle.buffer.rope();
-                        let bound = off.min(rope.len_bytes());
-                        *handle.buffer.cursors.primary_mut() =
-                            crate::buffer::cursor::Cursor::from_byte_offset(rope, bound);
-                        handle.buffer.cursors.collapse_to_primary();
-                    } else {
-                        self.status_error = Some(format!("No mark named {c}"));
-                    }
-                    return;
-                }
-                InputMode::RecordMacroChar => {
-                    self.input_mode = InputMode::Normal;
-                    self.macros.start_recording(c);
-                    return;
-                }
-                InputMode::ReplayMacroChar => {
-                    self.input_mode = InputMode::Normal;
-                    self.replay_macro_slot(c);
-                    return;
-                }
-                InputMode::SurroundChar => {
-                    self.input_mode = InputMode::Normal;
-                    self.apply_surround(c);
-                    return;
-                }
-                _ => {}
-            }
-        }
-        // Mutate the input string for typing/backspace without accessing other fields.
-        match action {
-            EditorAction::InsertChar(c) => {
-                // JumpToLine only accepts digits and `:`; every other
-                // string-carrying mode accepts any char.
-                let allow = !matches!(&self.input_mode, InputMode::JumpToLine(_))
-                    || c.is_ascii_digit()
-                    || c == ':';
-                if allow && let Some(s) = self.input_mode.string_mut() {
-                    s.push(c);
-                }
-                return;
-            }
-            EditorAction::DeleteBackward => {
-                if let Some(s) = self.input_mode.string_mut() {
-                    s.pop();
-                }
-                return;
-            }
-            _ => {}
-        }
-
-        // For Enter / Esc: take ownership of the mode to free the borrow before
-        // accessing self.editor or self.input_mode again.
-        match action {
-            EditorAction::InsertNewline => {
-                let mode = std::mem::replace(&mut self.input_mode, InputMode::Normal);
-                match mode {
-                    InputMode::JumpToLine(input) => {
-                        let (line_str, col_str) = match input.split_once(':') {
-                            Some((l, c)) => (l, Some(c)),
-                            None => (input.as_str(), None),
-                        };
-                        if let Ok(n) = line_str.parse::<usize>() {
-                            self.push_current_to_jump_list();
-                            let line = n.saturating_sub(1); // 1-based input
-                            let buf = &mut self.editor.active_mut().buffer;
-                            let target = {
-                                let rope = buf.rope();
-                                let clamped_line = line.min(rope.len_lines().saturating_sub(1));
-                                let line_start_char = rope.line_to_char(clamped_line);
-                                let target_char =
-                                    match col_str.and_then(|s| s.parse::<usize>().ok()) {
-                                        Some(col_n) => {
-                                            let col = col_n.saturating_sub(1); // 1-based → 0-based
-                                            let line_char_len = rope.line(clamped_line).len_chars();
-                                            // Exclude trailing newline when clamping the column.
-                                            let line_content = if line_char_len > 0
-                                                && rope.char(line_start_char + line_char_len - 1)
-                                                    == '\n'
-                                            {
-                                                line_char_len - 1
-                                            } else {
-                                                line_char_len
-                                            };
-                                            line_start_char + col.min(line_content)
-                                        }
-                                        None => line_start_char,
-                                    };
-                                rope.char_to_byte(target_char)
-                            };
-                            buf.move_cursor_to(target, false);
-                        }
-                    }
-                    InputMode::OpenFilePath(input) => {
-                        let path = PathBuf::from(input.trim());
-                        self.push_current_to_jump_list();
-                        let _ = self.editor.open_tab(path);
-                        self.after_file_open_or_save();
-                    }
-                    InputMode::SaveAsPath(input) => {
-                        let path = PathBuf::from(input.trim());
-                        let _ = self.editor.active_mut().save_as(path);
-                        self.after_file_open_or_save();
-                    }
-                    InputMode::RenamePath(original, input) => {
-                        let new_name = input.trim();
-                        // Validate: must be a plain filename (no path separators or ..).
-                        let mut components = std::path::Path::new(new_name).components();
-                        let is_plain_name = matches!(
-                            (components.next(), components.next()),
-                            (Some(std::path::Component::Normal(_)), None)
-                        );
-                        if is_plain_name && let Some(parent) = original.parent() {
-                            let new_path = parent.join(new_name);
-                            if !new_path.exists() && std::fs::rename(&original, &new_path).is_ok() {
-                                self.refresh_sidebar();
-                            }
-                        }
-                    }
-                    InputMode::NewFolderName(parent, input) => {
-                        let name = input.trim();
-                        let mut components = std::path::Path::new(name).components();
-                        let is_plain_name = matches!(
-                            (components.next(), components.next()),
-                            (Some(std::path::Component::Normal(_)), None)
-                        );
-                        if is_plain_name {
-                            let new_dir = parent.join(name);
-                            if !new_dir.exists() && std::fs::create_dir(&new_dir).is_ok() {
-                                self.refresh_sidebar();
-                            }
-                        }
-                    }
-                    InputMode::Rename(input) => {
-                        if !input.is_empty() {
-                            self.send_rename(&input);
-                        }
-                    }
-                    InputMode::GitCommitMessage(input) => {
-                        self.git_finish_commit(&input);
-                    }
-                    InputMode::GitNewBranch(input) => {
-                        self.git_finish_new_branch(&input);
-                    }
-                    InputMode::GitStashMessage(input) => {
-                        self.git_finish_stash_push(&input);
-                    }
-                    InputMode::ShellFilter(input) => {
-                        self.apply_shell_filter(&input);
-                    }
-                    InputMode::AlignChar(input) => {
-                        if let Some(c) = input.chars().next() {
-                            self.editor.active_mut().buffer.align_on(c);
-                        }
-                    }
-                    InputMode::Normal
-                    | InputMode::SetMarkChar
-                    | InputMode::JumpToMarkChar
-                    | InputMode::RecordMacroChar
-                    | InputMode::ReplayMacroChar
-                    | InputMode::SurroundChar => {}
-                }
-            }
-            EditorAction::Quit | EditorAction::Unhandled => {
-                self.input_mode = InputMode::Normal;
-            }
-            _ => {}
-        }
-    }
 
     // ── Fuzzy picker input handling ──────────────────────────────────────────
 
-    fn handle_fuzzy_picker(&mut self, action: EditorAction) {
-        if self.fuzzy_picker.is_none() {
-            return;
-        }
-        match action {
-            EditorAction::InsertChar(c) => {
-                if let Some(picker) = &mut self.fuzzy_picker {
-                    let mut q = picker.query.clone();
-                    q.push(c);
-                    picker.update_query(q);
-                }
-            }
-            EditorAction::DeleteBackward => {
-                if let Some(picker) = &mut self.fuzzy_picker {
-                    let mut q = picker.query.clone();
-                    q.pop();
-                    picker.update_query(q);
-                }
-            }
-            EditorAction::MoveCursor(Direction::Up) => {
-                if let Some(picker) = &mut self.fuzzy_picker {
-                    picker.move_up();
-                }
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                if let Some(picker) = &mut self.fuzzy_picker {
-                    picker.move_down();
-                }
-            }
-            EditorAction::InsertNewline => {
-                // Extract path before closing picker to avoid borrow conflict with self.editor.
-                let path = self
-                    .fuzzy_picker
-                    .as_ref()
-                    .and_then(|p| p.selected_path().cloned());
-                self.fuzzy_picker = None;
-                if let Some(path) = path {
-                    self.push_current_to_jump_list();
-                    let _ = self.editor.open_tab(path);
-                    self.after_file_open_or_save();
-                }
-            }
-            EditorAction::Quit | EditorAction::CloseSearch | EditorAction::Unhandled => {
-                self.fuzzy_picker = None;
-            }
-            _ => {}
-        }
-    }
-
-    fn handle_symbol_picker(&mut self, action: EditorAction) {
-        if self.symbol_picker.is_none() {
-            return;
-        }
-        match action {
-            EditorAction::InsertChar(c) => {
-                if let Some(picker) = &mut self.symbol_picker {
-                    let mut q = picker.query.clone();
-                    q.push(c);
-                    picker.update_query(q);
-                }
-            }
-            EditorAction::DeleteBackward => {
-                if let Some(picker) = &mut self.symbol_picker {
-                    let mut q = picker.query.clone();
-                    q.pop();
-                    picker.update_query(q);
-                }
-            }
-            EditorAction::MoveCursor(Direction::Up) => {
-                if let Some(picker) = &mut self.symbol_picker {
-                    picker.move_up();
-                }
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                if let Some(picker) = &mut self.symbol_picker {
-                    picker.move_down();
-                }
-            }
-            EditorAction::InsertNewline => {
-                let target = self
-                    .symbol_picker
-                    .as_ref()
-                    .and_then(|p| p.selected_symbol().cloned());
-                self.symbol_picker = None;
-                if let Some(sym) = target {
-                    self.push_current_to_jump_list();
-                    let handle = self.editor.active_mut();
-                    let rope = handle.buffer.rope();
-                    let bound = sym.byte_range.start.min(rope.len_bytes());
-                    *handle.buffer.cursors.primary_mut() =
-                        crate::buffer::cursor::Cursor::from_byte_offset(rope, bound);
-                    handle.buffer.cursors.collapse_to_primary();
-                }
-            }
-            EditorAction::Quit | EditorAction::CloseSearch | EditorAction::Unhandled => {
-                self.symbol_picker = None;
-            }
-            _ => {}
-        }
-    }
-
     // ── Project search overlay input handling ────────────────────────────────
-
-    /// Handle keyboard input while the project-search overlay is active.
-    /// Returns `true` when the action was consumed (do not fall through to the
-    /// editor). Global actions like Quit/ToggleHelp fall through.
-    fn handle_project_search(&mut self, action: EditorAction) -> bool {
-        match action {
-            EditorAction::InsertChar(c) => {
-                if let Some(ps) = &mut self.project_search {
-                    if ps.focus_replace {
-                        ps.replace_text.push(c);
-                    } else {
-                        ps.query.push(c);
-                    }
-                }
-                if !self
-                    .project_search
-                    .as_ref()
-                    .map(|s| s.focus_replace)
-                    .unwrap_or(false)
-                {
-                    self.recompute_project_search();
-                }
-                true
-            }
-            EditorAction::DeleteBackward => {
-                if let Some(ps) = &mut self.project_search {
-                    if ps.focus_replace {
-                        ps.replace_text.pop();
-                    } else {
-                        ps.query.pop();
-                    }
-                }
-                if !self
-                    .project_search
-                    .as_ref()
-                    .map(|s| s.focus_replace)
-                    .unwrap_or(false)
-                {
-                    self.recompute_project_search();
-                }
-                true
-            }
-            EditorAction::MoveCursor(Direction::Up) => {
-                if let Some(ps) = &mut self.project_search {
-                    ps.move_up();
-                }
-                true
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                if let Some(ps) = &mut self.project_search {
-                    ps.move_down();
-                }
-                true
-            }
-            EditorAction::InsertTab => {
-                // Tab toggles focus and reveals the replace input.
-                if let Some(ps) = &mut self.project_search {
-                    ps.show_replace = true;
-                    ps.focus_replace = !ps.focus_replace;
-                }
-                true
-            }
-            EditorAction::SearchToggleRegex => {
-                if let Some(ps) = &mut self.project_search {
-                    ps.is_regex = !ps.is_regex;
-                }
-                self.recompute_project_search();
-                true
-            }
-            EditorAction::SearchToggleCaseSensitive => {
-                if let Some(ps) = &mut self.project_search {
-                    ps.case_sensitive = !ps.case_sensitive;
-                }
-                self.recompute_project_search();
-                true
-            }
-            EditorAction::InsertNewline => {
-                let target = self
-                    .project_search
-                    .as_ref()
-                    .and_then(|ps| ps.results.matches.get(ps.selected).cloned());
-                if let Some(m) = target {
-                    let abs = self.workspace.join(&m.path);
-                    self.project_search = None;
-                    let _ = self.editor.open_tab(abs);
-                    self.after_file_open_or_save();
-                    // Move cursor to the start of the match line. Computing the
-                    // exact column requires the file's full byte→line table,
-                    // which we don't cache; landing on the line is the
-                    // standard "go to result" behaviour.
-                    let buf = &mut self.editor.active_mut().buffer;
-                    let rope = buf.rope();
-                    let line_idx = m.line.min(rope.len_lines().saturating_sub(1));
-                    let line_start_char = rope.line_to_char(line_idx);
-                    let target_byte = rope.char_to_byte(line_start_char);
-                    buf.move_cursor_to(target_byte, false);
-                }
-                true
-            }
-            EditorAction::SearchReplaceAll => {
-                self.project_replace_all();
-                true
-            }
-            EditorAction::CloseSearch | EditorAction::Quit | EditorAction::Unhandled => {
-                self.project_search = None;
-                true
-            }
-            // Global actions that should still work while the overlay is open
-            // (toggle help, scroll cursor, etc.).
-            EditorAction::ToggleHelp => false,
-            _ => true,
-        }
-    }
-
-    fn recompute_project_search(&mut self) {
-        let (query, is_regex, case_sensitive) = match &self.project_search {
-            Some(ps) => (ps.query.clone(), ps.is_regex, ps.case_sensitive),
-            None => return,
-        };
-        let results = crate::search::project::run(
-            &self.workspace,
-            &query,
-            is_regex,
-            case_sensitive,
-            self.config.hide_git_folder,
-            self.config.hide_dot_folders,
-        );
-        if let Some(ps) = &mut self.project_search {
-            ps.results = results;
-            ps.selected = 0;
-        }
-    }
-
-    fn project_replace_all(&mut self) {
-        let (query, replacement, is_regex, case_sensitive) = match &self.project_search {
-            Some(ps) if !ps.query.is_empty() && ps.show_replace => (
-                ps.query.clone(),
-                ps.replace_text.clone(),
-                ps.is_regex,
-                ps.case_sensitive,
-            ),
-            _ => return,
-        };
-
-        // Collect distinct file paths from results.
-        let paths: Vec<PathBuf> = {
-            let ps = match &self.project_search {
-                Some(p) => p,
-                None => return,
-            };
-            let mut seen = std::collections::HashSet::new();
-            ps.results
-                .matches
-                .iter()
-                .filter_map(|m| {
-                    let abs = self.workspace.join(&m.path);
-                    if seen.insert(abs.clone()) {
-                        Some(abs)
-                    } else {
-                        None
-                    }
-                })
-                .collect()
-        };
-
-        let mut total_replaced = 0usize;
-        for path in &paths {
-            // If the file is currently open in a tab, edit it through the
-            // buffer so undo history is preserved.
-            let open_idx = self
-                .editor
-                .tabs
-                .iter()
-                .position(|t| t.path.as_deref() == Some(path.as_path()));
-            if let Some(idx) = open_idx {
-                let prev_active = self.editor.active_idx;
-                self.editor.go_to_tab(idx);
-                let buf = &mut self.editor.active_mut().buffer;
-                let text = buf.to_string();
-                let pattern = crate::search::build_pattern(&query, is_regex, case_sensitive);
-                let re = match regex::Regex::new(&pattern) {
-                    Ok(r) => r,
-                    Err(_) => continue,
-                };
-                let ranges: Vec<(usize, usize)> =
-                    re.find_iter(&text).map(|m| (m.start(), m.end())).collect();
-                if ranges.is_empty() {
-                    self.editor.go_to_tab(prev_active);
-                    continue;
-                }
-                buf.begin_batch();
-                for (start, end) in ranges.iter().rev() {
-                    buf.move_cursor_to(*start, false);
-                    buf.move_cursor_to(*end, true);
-                    buf.insert_str(&replacement);
-                }
-                buf.commit_batch();
-                total_replaced += ranges.len();
-                self.editor.go_to_tab(prev_active);
-            } else {
-                // Edit on disk.
-                if let Ok(n) = crate::search::project::replace_all_in_file(
-                    path,
-                    &query,
-                    is_regex,
-                    case_sensitive,
-                    &replacement,
-                ) {
-                    total_replaced += n;
-                }
-            }
-        }
-
-        // Refresh result list and surface a status message.
-        self.recompute_project_search();
-        self.status_error = Some(format!("Replaced {total_replaced} occurrence(s)"));
-    }
 
     // ── Search input handling ────────────────────────────────────────────────
 
-    /// Handle keyboard input while the search bar is active.
-    /// Returns `true` if the action was consumed (should not be processed further).
-    fn handle_search_input(&mut self, action: EditorAction) -> bool {
-        match &action {
-            // Navigation: let these fall through to the normal dispatch.
-            EditorAction::MoveCursor(_)
-            | EditorAction::MoveCursorWord(_)
-            | EditorAction::MoveCursorHome
-            | EditorAction::MoveCursorEnd
-            | EditorAction::MoveCursorFileStart
-            | EditorAction::MoveCursorFileEnd
-            | EditorAction::MoveCursorPage(_)
-            | EditorAction::Scroll(_)
-            | EditorAction::ScrollCursorCenter
-            | EditorAction::GoToMatchingBracket
-            | EditorAction::MouseClick { .. }
-            | EditorAction::MouseDrag { .. } => return false,
-
-            _ => {}
-        }
-
-        match action {
-            EditorAction::InsertChar(c) => {
-                let focus_replace = self
-                    .search_state
-                    .as_ref()
-                    .map(|s| s.focus_replace)
-                    .unwrap_or(false);
-                if focus_replace {
-                    if let Some(ss) = &mut self.search_state {
-                        ss.replace_text.push(c);
-                    }
-                } else {
-                    if let Some(ss) = &mut self.search_state {
-                        ss.query.push(c);
-                    }
-                    self.recompute_search_and_jump();
-                }
-            }
-            EditorAction::DeleteBackward => {
-                let focus_replace = self
-                    .search_state
-                    .as_ref()
-                    .map(|s| s.focus_replace)
-                    .unwrap_or(false);
-                if focus_replace {
-                    if let Some(ss) = &mut self.search_state {
-                        ss.replace_text.pop();
-                    }
-                } else {
-                    if let Some(ss) = &mut self.search_state {
-                        ss.query.pop();
-                    }
-                    self.recompute_search_and_jump();
-                }
-            }
-            EditorAction::InsertNewline => {
-                let focus_replace = self
-                    .search_state
-                    .as_ref()
-                    .map(|s| s.focus_replace && s.show_replace)
-                    .unwrap_or(false);
-                if focus_replace {
-                    self.replace_current();
-                } else {
-                    self.search_next();
-                }
-            }
-            EditorAction::InsertTab => {
-                // Toggle focus between query and replace fields.
-                if let Some(ss) = &mut self.search_state
-                    && ss.show_replace
-                {
-                    ss.focus_replace = !ss.focus_replace;
-                }
-            }
-            EditorAction::SearchNext => self.search_next(),
-            EditorAction::SearchPrev => self.search_prev(),
-            EditorAction::SearchReplaceOne => self.replace_current(),
-            EditorAction::SearchReplaceAll => self.replace_all(),
-            EditorAction::SearchToggleRegex => {
-                if let Some(ss) = &mut self.search_state {
-                    ss.is_regex = !ss.is_regex;
-                }
-                self.recompute_search_and_jump();
-            }
-            EditorAction::SearchToggleCaseSensitive => {
-                if let Some(ss) = &mut self.search_state {
-                    ss.case_sensitive = !ss.case_sensitive;
-                }
-                self.recompute_search_and_jump();
-            }
-            EditorAction::OpenReplace => {
-                if let Some(ss) = &mut self.search_state {
-                    ss.show_replace = true;
-                    ss.focus_replace = true;
-                }
-            }
-            EditorAction::CloseSearch | EditorAction::Quit => {
-                self.search_state = None;
-            }
-            EditorAction::ToggleHelp => {
-                self.show_help = !self.show_help;
-                if self.show_help {
-                    self.help_scroll = 0;
-                }
-            }
-            _ => {}
-        }
-        true
-    }
-
     // ── Search helpers ────────────────────────────────────────────────────────
-
-    fn recompute_search_and_jump(&mut self) {
-        if self.search_state.is_none() {
-            return;
-        }
-        let text = self.editor.active().buffer.to_string();
-        let cursor_offset = self.editor.active().buffer.cursors.primary().byte_offset;
-        if let Some(ss) = &mut self.search_state {
-            ss.recompute_matches(&text);
-            ss.jump_to_nearest(cursor_offset);
-        }
-        self.select_current_match();
-    }
-
-    fn search_next(&mut self) {
-        if let Some(ss) = &mut self.search_state {
-            ss.next_match();
-        }
-        self.select_current_match();
-    }
-
-    fn search_prev(&mut self) {
-        if let Some(ss) = &mut self.search_state {
-            ss.prev_match();
-        }
-        self.select_current_match();
-    }
-
-    fn select_current_match(&mut self) {
-        let range = self.search_state.as_ref().and_then(|s| s.current_range());
-        if let Some(r) = range {
-            self.editor
-                .active_mut()
-                .buffer
-                .move_cursor_to(r.start, false);
-            self.editor.active_mut().buffer.move_cursor_to(r.end, true);
-        }
-    }
-
-    fn replace_current(&mut self) {
-        let range = self.search_state.as_ref().and_then(|s| s.current_range());
-        let replace_text = self
-            .search_state
-            .as_ref()
-            .map(|s| s.replace_text.clone())
-            .unwrap_or_default();
-        if let Some(r) = range {
-            let buf = &mut self.editor.active_mut().buffer;
-            buf.begin_batch();
-            buf.move_cursor_to(r.start, false);
-            buf.move_cursor_to(r.end, true);
-            buf.insert_str(&replace_text);
-            buf.commit_batch();
-        }
-        self.recompute_search_and_jump();
-    }
-
-    fn replace_all(&mut self) {
-        let ranges: Vec<_> = self
-            .search_state
-            .as_ref()
-            .map(|s| s.matches.clone())
-            .unwrap_or_default();
-        let replace_text = self
-            .search_state
-            .as_ref()
-            .map(|s| s.replace_text.clone())
-            .unwrap_or_default();
-        if ranges.is_empty() {
-            return;
-        }
-
-        let buf = &mut self.editor.active_mut().buffer;
-        buf.begin_batch();
-        // Replace in reverse order so earlier byte offsets remain valid.
-        for r in ranges.iter().rev() {
-            buf.move_cursor_to(r.start, false);
-            buf.move_cursor_to(r.end, true);
-            buf.insert_str(&replace_text);
-        }
-        buf.commit_batch();
-
-        self.recompute_search_and_jump();
-    }
-
-    fn select_all_occurrences(&mut self) {
-        // Use the current selection text as the search query (or keep existing query).
-        if let Some(selected) = self.selected_text() {
-            if self.search_state.is_none() {
-                self.search_state = Some(SearchState::new(false));
-            }
-            if let Some(ss) = &mut self.search_state {
-                ss.query = selected;
-                ss.case_sensitive = true;
-            }
-        }
-        self.recompute_search_and_jump();
-    }
 
     // ── Help overlay input handling ───────────────────────────────────────────
 
-    /// Persist `last_seen_version = current` so the same overlay isn't shown
-    /// again until the next minor/major upgrade.
-    fn record_version_seen(&mut self) {
-        let v = env!("CARGO_PKG_VERSION").to_string();
-        if self.config.last_seen_version.as_deref() != Some(v.as_str()) {
-            self.config.last_seen_version = Some(v);
-            self.config.save();
-        }
-    }
-
-    /// Handle input while the welcome overlay is visible. Any "OK" key
-    /// (Enter / Esc / F1 / Ctrl+Q-style ToggleHelp) dismisses it; arrows and
-    /// the mouse wheel scroll its content. Returns `true` to consume.
-    fn handle_welcome(&mut self, action: &EditorAction) -> bool {
-        if scroll_action(action, &mut self.welcome_scroll) {
-            return true;
-        }
-        match action {
-            EditorAction::InsertNewline | EditorAction::CloseSearch => {
-                self.show_welcome = false;
-                self.record_version_seen();
-                true
-            }
-            // Don't let typing leak through to the editor while the modal is up.
-            EditorAction::InsertChar(_) | EditorAction::InsertTab => true,
-            _ => false,
-        }
-    }
-
-    /// Handle input while the changelog overlay is visible. Same dismiss /
-    /// scroll semantics as the welcome overlay.
-    fn handle_changelog(&mut self, action: &EditorAction) -> bool {
-        if scroll_action(action, &mut self.changelog_scroll) {
-            return true;
-        }
-        match action {
-            EditorAction::InsertNewline | EditorAction::CloseSearch => {
-                self.show_changelog = false;
-                self.record_version_seen();
-                true
-            }
-            EditorAction::InsertChar(_) | EditorAction::InsertTab => true,
-            _ => false,
-        }
-    }
-
-    /// Handle input while the help overlay is visible.
-    /// Returns `true` if the action was consumed (caller should `return`).
-    fn handle_help(&mut self, action: &EditorAction) -> bool {
-        if scroll_action(action, &mut self.help_scroll) {
-            return true;
-        }
-        match action {
-            EditorAction::MoveCursorFileStart => {
-                self.help_scroll = 0;
-                true
-            }
-            EditorAction::MoveCursorFileEnd => {
-                self.help_scroll = usize::MAX; // clamped in render
-                true
-            }
-            EditorAction::ToggleHelp | EditorAction::CloseSearch => {
-                self.show_help = false;
-                true
-            }
-            // Swallow all text-insertion actions so they don't reach the editor.
-            EditorAction::InsertChar(_) | EditorAction::InsertNewline | EditorAction::InsertTab => {
-                true
-            }
-            _ => false,
-        }
-    }
-
     // ── Settings overlay input handling ──────────────────────────────────────
-
-    /// Handle input while the settings overlay is open.
-    /// Returns `true` if the action was consumed, `false` to let it fall through.
-    fn handle_settings(&mut self, action: &EditorAction) -> bool {
-        const NUM_ROWS: usize = crate::ui::settings_overlay::NUM_SETTINGS;
-        match action {
-            EditorAction::MoveCursor(Direction::Up) => {
-                self.settings_cursor = self.settings_cursor.saturating_sub(1);
-                true
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                self.settings_cursor = (self.settings_cursor + 1).min(NUM_ROWS - 1);
-                true
-            }
-            EditorAction::InsertChar(' ') | EditorAction::InsertNewline => {
-                self.toggle_setting(true);
-                true
-            }
-            EditorAction::MoveCursor(Direction::Right) => {
-                self.toggle_setting(true);
-                true
-            }
-            EditorAction::MoveCursor(Direction::Left) => {
-                self.toggle_setting(false);
-                true
-            }
-            // Let Quit / Escape close the overlay but fall through so Quit still quits.
-            EditorAction::OpenSettings | EditorAction::CloseSearch => {
-                self.show_settings = false;
-                true
-            }
-            EditorAction::Quit | EditorAction::ForceQuit => {
-                self.show_settings = false;
-                false
-            }
-            // Swallow all text-insertion actions so they don't reach the editor.
-            EditorAction::InsertChar(_) | EditorAction::InsertTab => true,
-            _ => false,
-        }
-    }
-
-    /// Toggle or cycle the setting at `settings_cursor`. `forward` controls
-    /// direction for enum settings; booleans always flip.
-    fn toggle_setting(&mut self, forward: bool) {
-        match self.settings_cursor {
-            0 => self.config.confirm_exit = !self.config.confirm_exit,
-            1 => self.config.auto_save = !self.config.auto_save,
-            2 => self.config.show_whitespace = !self.config.show_whitespace,
-            3 => {
-                self.config.highlight_trailing_whitespace =
-                    !self.config.highlight_trailing_whitespace;
-            }
-            4 => self.config.warn_mixed_indent = !self.config.warn_mixed_indent,
-            5 => self.config.auto_pair = !self.config.auto_pair,
-            6 => self.config.hide_git_folder = !self.config.hide_git_folder,
-            7 => self.config.hide_dot_folders = !self.config.hide_dot_folders,
-            8 => self.config.restore_session = !self.config.restore_session,
-            9 => self.config.persistent_undo = !self.config.persistent_undo,
-            10 => {
-                let all = Theme::ALL;
-                let idx = all
-                    .iter()
-                    .position(|t| t == &self.config.theme)
-                    .unwrap_or(0);
-                let next = if forward {
-                    (idx + 1) % all.len()
-                } else {
-                    (idx + all.len() - 1) % all.len()
-                };
-                self.config.theme = all[next].clone();
-            }
-            11 => {
-                let all = KeymapPreset::ALL;
-                let idx = all
-                    .iter()
-                    .position(|p| p == &self.config.keymap_preset)
-                    .unwrap_or(0);
-                let next = if forward {
-                    (idx + 1) % all.len()
-                } else {
-                    (idx + all.len() - 1) % all.len()
-                };
-                self.config.keymap_preset = all[next].clone();
-                KeyBindings::apply_preset(&self.config.keymap_preset);
-                self.input.reload_keybindings();
-            }
-            _ => {}
-        }
-        self.config.save();
-    }
 
     // ── LSP picker input handling ────────────────────────────────────────────
 
-    /// Handle input while the LSP config picker is open.
-    /// Returns `true` if the action was consumed, `false` to let it fall through.
-    fn handle_lsp_picker(&mut self, action: &EditorAction) -> bool {
-        let num_rows = 1 + LSP_SERVER_OPTIONS.len();
-        match action {
-            EditorAction::MoveCursor(Direction::Up) => {
-                if let Some(picker) = &mut self.lsp_picker {
-                    picker.selected = picker.selected.saturating_sub(1);
-                }
-                true
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                if let Some(picker) = &mut self.lsp_picker {
-                    picker.selected = (picker.selected + 1).min(num_rows - 1);
-                }
-                true
-            }
-            EditorAction::InsertChar(' ') | EditorAction::InsertNewline => {
-                self.apply_lsp_picker_selection();
-                self.lsp_picker = None;
-                true
-            }
-            EditorAction::OpenLspConfig | EditorAction::CloseSearch => {
-                self.lsp_picker = None;
-                true
-            }
-            EditorAction::Quit | EditorAction::ForceQuit => {
-                self.lsp_picker = None;
-                false
-            }
-            _ => false,
-        }
-    }
-
-    /// Write the selected LSP config to `<workspace>/.txt/lsp.toml` and reload.
-    fn apply_lsp_picker_selection(&mut self) {
-        let selected = match &self.lsp_picker {
-            Some(p) => p.selected,
-            None => return,
-        };
-
-        use crate::lsp::config::{LspServerEntry, WorkspaceLspConfig};
-        use std::collections::HashMap;
-
-        let new_config = if selected == 0 {
-            // Disabled
-            WorkspaceLspConfig::default()
-        } else {
-            let (_language, name, command, args) = LSP_SERVER_OPTIONS[selected - 1];
-            let mut servers = HashMap::new();
-            servers.insert(
-                name.to_string(),
-                LspServerEntry {
-                    command: command.to_string(),
-                    args: args.iter().map(|s| s.to_string()).collect(),
-                    init_options: None,
-                },
-            );
-            WorkspaceLspConfig {
-                enabled: true,
-                server: Some(name.to_string()),
-                servers,
-            }
-        };
-
-        // Write config file.
-        let txt_dir = self.workspace.join(".txt");
-        let _ = std::fs::create_dir_all(&txt_dir);
-        if let Ok(text) = toml::to_string(&new_config) {
-            let _ = std::fs::write(txt_dir.join("lsp.toml"), text);
-        }
-
-        // Tear down existing LSP connection if any.
-        self.lsp = None;
-        self.pending_lsp_approval = None;
-
-        // Apply new config.
-        self.lsp_config = new_config;
-
-        // Start new server if enabled — routed through the trust gate.
-        self.request_lsp_start();
-    }
-
     // ── Git operations dialog ────────────────────────────────────────────────
-
-    fn open_git_dialog(&mut self) {
-        if !crate::git::ops::is_repo(&self.workspace) {
-            self.status_error = Some("Not a git repository".into());
-            return;
-        }
-        self.git_dialog = Some(GitDialogState::new());
-    }
-
-    /// Drive the git dialog. Captures all input until `Esc` from the menu
-    /// closes the overlay. Routes nav keys to `GitScreen::move_*`, and
-    /// handles per-screen actions inline.
-    fn handle_git_dialog(&mut self, action: EditorAction) {
-        use crate::ui::git_dialog::{ConfirmOp, GitScreen, MenuItem};
-
-        // Quit always closes the dialog without doing anything.
-        if matches!(action, EditorAction::Quit | EditorAction::ForceQuit) {
-            self.git_dialog = None;
-            return;
-        }
-
-        // Esc / CloseSearch: step back; if no history, close the dialog.
-        if matches!(action, EditorAction::CloseSearch) {
-            let close = match self.git_dialog.as_mut() {
-                Some(d) => !d.pop(),
-                None => true,
-            };
-            if close {
-                self.git_dialog = None;
-            }
-            return;
-        }
-
-        // Navigation keys are uniform across most screens.
-        match action {
-            EditorAction::MoveCursor(Direction::Up) => {
-                if let Some(d) = self.git_dialog.as_mut() {
-                    d.screen.move_up();
-                    d.screen.scroll_by(-1);
-                }
-                return;
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                if let Some(d) = self.git_dialog.as_mut() {
-                    d.screen.move_down();
-                    d.screen.scroll_by(1);
-                }
-                return;
-            }
-            EditorAction::MoveCursorPage(Direction::Up) | EditorAction::Scroll(ScrollDir::Up) => {
-                if let Some(d) = self.git_dialog.as_mut() {
-                    d.screen.scroll_by(-5);
-                }
-                return;
-            }
-            EditorAction::MoveCursorPage(Direction::Down)
-            | EditorAction::Scroll(ScrollDir::Down) => {
-                if let Some(d) = self.git_dialog.as_mut() {
-                    d.screen.scroll_by(5);
-                }
-                return;
-            }
-            _ => {}
-        }
-
-        // Per-screen handling. We snapshot the current screen kind so we can
-        // borrow self mutably below for I/O without holding a borrow on
-        // `self.git_dialog`.
-        let screen_kind = match self.git_dialog.as_ref() {
-            Some(d) => d.screen.clone(),
-            None => return,
-        };
-
-        match (screen_kind, action) {
-            // ── Menu ──
-            (GitScreen::Menu { selected }, EditorAction::InsertNewline) => {
-                let item = MenuItem::ALL[selected];
-                self.git_open_menu_item(item);
-            }
-
-            // ── Stage ──
-            (
-                GitScreen::Stage { entries, .. },
-                EditorAction::InsertChar(' ') | EditorAction::InsertChar('x'),
-            ) => {
-                if let Some(GitDialogState {
-                    screen:
-                        GitScreen::Stage {
-                            checked, selected, ..
-                        },
-                    ..
-                }) = self.git_dialog.as_mut()
-                    && let Some(slot) = checked.get_mut(*selected)
-                    && !entries.is_empty()
-                {
-                    *slot = !*slot;
-                }
-            }
-            (
-                GitScreen::Stage {
-                    entries, checked, ..
-                },
-                EditorAction::InsertNewline,
-            ) => {
-                self.git_apply_stage(&entries, &checked);
-            }
-
-            // ── Branches ──
-            (GitScreen::Branches { entries, selected }, EditorAction::InsertNewline) => {
-                if let Some(branch) = entries.get(selected) {
-                    if branch.current {
-                        if let Some(d) = self.git_dialog.as_mut() {
-                            d.set_error("Already on this branch");
-                        }
-                    } else {
-                        let name = branch.name.clone();
-                        match crate::git::ops::checkout(&self.workspace, &name) {
-                            Ok(out) => {
-                                self.git_after_branch_change();
-                                self.set_git_output("Checkout", out);
-                            }
-                            Err(e) => self.set_git_error(e),
-                        }
-                    }
-                }
-            }
-            (GitScreen::Branches { .. }, EditorAction::InsertChar('n')) => {
-                self.input_mode = InputMode::GitNewBranch(String::new());
-            }
-            (GitScreen::Branches { entries, selected }, EditorAction::InsertChar('d')) => {
-                if let Some(branch) = entries.get(selected) {
-                    if branch.current {
-                        if let Some(d) = self.git_dialog.as_mut() {
-                            d.set_error("Cannot delete the current branch");
-                        }
-                    } else if let Some(d) = self.git_dialog.as_mut() {
-                        let op = ConfirmOp::DeleteBranch(branch.name.clone());
-                        d.push(GitScreen::Confirm { op });
-                    }
-                }
-            }
-
-            // ── Stashes ──
-            (GitScreen::Stashes { entries, selected }, EditorAction::InsertNewline) => {
-                if let Some(s) = entries.get(selected) {
-                    let idx = s.index;
-                    match crate::git::ops::stash_apply(&self.workspace, idx) {
-                        Ok(out) => self.set_git_output("Stash apply", out),
-                        Err(e) => self.set_git_error(e),
-                    }
-                }
-            }
-            (GitScreen::Stashes { entries, selected }, EditorAction::InsertChar('p')) => {
-                if let Some(s) = entries.get(selected) {
-                    let idx = s.index;
-                    match crate::git::ops::stash_pop(&self.workspace, idx) {
-                        Ok(out) => {
-                            self.git_after_branch_change();
-                            self.set_git_output("Stash pop", out);
-                        }
-                        Err(e) => self.set_git_error(e),
-                    }
-                }
-            }
-            (GitScreen::Stashes { entries, selected }, EditorAction::InsertChar('d')) => {
-                if let Some(s) = entries.get(selected)
-                    && let Some(d) = self.git_dialog.as_mut()
-                {
-                    let op = ConfirmOp::DropStash(s.index);
-                    d.push(GitScreen::Confirm { op });
-                }
-            }
-            (GitScreen::Stashes { .. }, EditorAction::InsertChar('n')) => {
-                self.input_mode = InputMode::GitStashMessage(String::new());
-            }
-
-            // ── Confirm (y/n) ──
-            (GitScreen::Confirm { op }, EditorAction::InsertChar('y' | 'Y')) => {
-                self.git_run_confirm(op);
-            }
-            (GitScreen::Confirm { .. }, EditorAction::InsertChar('n' | 'N')) => {
-                if let Some(d) = self.git_dialog.as_mut() {
-                    d.pop();
-                }
-            }
-
-            _ => {}
-        }
-    }
-
-    fn git_open_menu_item(&mut self, item: crate::ui::git_dialog::MenuItem) {
-        use crate::ui::git_dialog::{GitScreen, MenuItem};
-
-        match item {
-            MenuItem::Status => match crate::git::ops::status_summary(&self.workspace) {
-                Ok(out) => {
-                    if let Some(d) = self.git_dialog.as_mut() {
-                        d.push(GitScreen::Status {
-                            output: out,
-                            scroll: 0,
-                        });
-                    }
-                }
-                Err(e) => self.set_git_error(e),
-            },
-            MenuItem::Stage => match crate::git::ops::status(&self.workspace) {
-                Ok(entries) => {
-                    if let Some(d) = self.git_dialog.as_mut() {
-                        let checked = vec![false; entries.len()];
-                        d.push(GitScreen::Stage {
-                            entries,
-                            checked,
-                            selected: 0,
-                        });
-                    }
-                }
-                Err(e) => self.set_git_error(e),
-            },
-            MenuItem::Commit => {
-                self.input_mode = InputMode::GitCommitMessage(String::new());
-            }
-            MenuItem::Push => match crate::git::ops::push(&self.workspace) {
-                Ok(out) => self.set_git_output("Push", out),
-                Err(e) => self.set_git_error(e),
-            },
-            MenuItem::Pull => match crate::git::ops::pull(&self.workspace) {
-                Ok(out) => {
-                    self.git_after_branch_change();
-                    self.set_git_output("Pull", out);
-                }
-                Err(e) => self.set_git_error(e),
-            },
-            MenuItem::Branches => match crate::git::ops::branches(&self.workspace) {
-                Ok(entries) => {
-                    if let Some(d) = self.git_dialog.as_mut() {
-                        let selected = entries.iter().position(|b| b.current).unwrap_or(0);
-                        d.push(GitScreen::Branches { entries, selected });
-                    }
-                }
-                Err(e) => self.set_git_error(e),
-            },
-            MenuItem::Stashes => match crate::git::ops::stashes(&self.workspace) {
-                Ok(entries) => {
-                    if let Some(d) = self.git_dialog.as_mut() {
-                        d.push(GitScreen::Stashes {
-                            entries,
-                            selected: 0,
-                        });
-                    }
-                }
-                Err(e) => self.set_git_error(e),
-            },
-        }
-    }
-
-    /// Apply staging based on the user's checked/unchecked selections.
-    /// Files that are currently staged become `git reset` targets; files that
-    /// are unstaged or untracked become `git add` targets.
-    fn git_apply_stage(&mut self, entries: &[crate::git::ops::StatusEntry], checked: &[bool]) {
-        let mut to_add: Vec<&std::path::Path> = Vec::new();
-        let mut to_reset: Vec<&std::path::Path> = Vec::new();
-        for (entry, &is_checked) in entries.iter().zip(checked.iter()) {
-            if !is_checked {
-                continue;
-            }
-            if entry.is_staged() {
-                to_reset.push(&entry.path);
-            } else {
-                to_add.push(&entry.path);
-            }
-        }
-
-        if to_add.is_empty() && to_reset.is_empty() {
-            if let Some(d) = self.git_dialog.as_mut() {
-                d.set_error("Nothing selected");
-            }
-            return;
-        }
-
-        if let Err(e) = crate::git::ops::add(&self.workspace, &to_add) {
-            self.set_git_error(e);
-            return;
-        }
-        if let Err(e) = crate::git::ops::reset(&self.workspace, &to_reset) {
-            self.set_git_error(e);
-            return;
-        }
-
-        // Refresh the stage screen with new statuses.
-        match crate::git::ops::status(&self.workspace) {
-            Ok(entries) => {
-                if let Some(d) = self.git_dialog.as_mut() {
-                    let checked = vec![false; entries.len()];
-                    d.replace(crate::ui::git_dialog::GitScreen::Stage {
-                        entries,
-                        checked,
-                        selected: 0,
-                    });
-                }
-            }
-            Err(e) => self.set_git_error(e),
-        }
-        self.refresh_git_gutter();
-    }
-
-    fn git_finish_commit(&mut self, message: &str) {
-        let trimmed = message.trim();
-        if trimmed.is_empty() {
-            if let Some(d) = self.git_dialog.as_mut() {
-                d.set_error("Empty commit message — cancelled");
-            }
-            return;
-        }
-        match crate::git::ops::commit(&self.workspace, trimmed) {
-            Ok(out) => {
-                self.set_git_output("Commit", out);
-                self.refresh_git_gutter();
-            }
-            Err(e) => self.set_git_error(e),
-        }
-    }
-
-    fn git_finish_new_branch(&mut self, name: &str) {
-        let trimmed = name.trim();
-        if trimmed.is_empty() {
-            if let Some(d) = self.git_dialog.as_mut() {
-                d.set_error("Empty branch name — cancelled");
-            }
-            return;
-        }
-        match crate::git::ops::create_branch(&self.workspace, trimmed) {
-            Ok(out) => {
-                self.git_after_branch_change();
-                self.set_git_output("Create branch", out);
-            }
-            Err(e) => self.set_git_error(e),
-        }
-    }
-
-    fn git_finish_stash_push(&mut self, message: &str) {
-        let trimmed = message.trim();
-        let msg = if trimmed.is_empty() {
-            None
-        } else {
-            Some(trimmed)
-        };
-        match crate::git::ops::stash_push(&self.workspace, msg) {
-            Ok(out) => {
-                self.git_after_branch_change();
-                self.set_git_output("Stash push", out);
-            }
-            Err(e) => self.set_git_error(e),
-        }
-    }
-
-    fn git_run_confirm(&mut self, op: crate::ui::git_dialog::ConfirmOp) {
-        use crate::ui::git_dialog::ConfirmOp;
-        // Pop the confirm screen first so the result lands on the prior screen.
-        if let Some(d) = self.git_dialog.as_mut() {
-            d.pop();
-        }
-        match op {
-            ConfirmOp::DropStash(idx) => match crate::git::ops::stash_drop(&self.workspace, idx) {
-                Ok(out) => {
-                    // Refresh the stash list under us.
-                    if let Ok(entries) = crate::git::ops::stashes(&self.workspace)
-                        && let Some(d) = self.git_dialog.as_mut()
-                    {
-                        d.replace(crate::ui::git_dialog::GitScreen::Stashes {
-                            entries,
-                            selected: 0,
-                        });
-                    }
-                    self.set_git_output("Stash drop", out);
-                }
-                Err(e) => self.set_git_error(e),
-            },
-            ConfirmOp::DeleteBranch(name) => {
-                match crate::git::ops::delete_branch(&self.workspace, &name) {
-                    Ok(out) => {
-                        if let Ok(entries) = crate::git::ops::branches(&self.workspace)
-                            && let Some(d) = self.git_dialog.as_mut()
-                        {
-                            let selected = entries.iter().position(|b| b.current).unwrap_or(0);
-                            d.replace(crate::ui::git_dialog::GitScreen::Branches {
-                                entries,
-                                selected,
-                            });
-                        }
-                        self.set_git_output("Delete branch", out);
-                    }
-                    Err(e) => self.set_git_error(e),
-                }
-            }
-        }
-    }
-
-    fn set_git_output(&mut self, title: &str, body: String) {
-        if let Some(d) = self.git_dialog.as_mut() {
-            d.push(crate::ui::git_dialog::GitScreen::Output {
-                title: title.into(),
-                body,
-                scroll: 0,
-            });
-        }
-    }
-
-    fn set_git_error(&mut self, err: String) {
-        if let Some(d) = self.git_dialog.as_mut() {
-            d.set_error(err);
-        }
-    }
-
-    /// Called after operations that may change which file is on disk under
-    /// the active buffer (checkout, pull, stash pop, new branch). Refreshes
-    /// the gutter; the existing file watcher will pick up content changes.
-    fn git_after_branch_change(&mut self) {
-        self.refresh_git_gutter();
-    }
 
     // ── Sidebar input handling ────────────────────────────────────────────────
 
-    /// Handle input while the sidebar is focused.
-    /// Returns `true` if the action was consumed, `false` to let it fall through.
-    fn handle_sidebar_input(&mut self, action: &EditorAction) -> bool {
-        match action {
-            EditorAction::MoveCursor(Direction::Up) => {
-                if let Some(sb) = &mut self.sidebar {
-                    sb.move_up();
-                }
-                self.ensure_sidebar_selected_visible();
-                true
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                if let Some(sb) = &mut self.sidebar {
-                    sb.move_down();
-                }
-                self.ensure_sidebar_selected_visible();
-                true
-            }
-            EditorAction::InsertNewline => {
-                // Enter: open file or expand/collapse directory.
-                let selected_path = self
-                    .sidebar
-                    .as_ref()
-                    .and_then(|sb| sb.selected_path().cloned());
-                if let Some(path) = selected_path {
-                    if path.is_dir() {
-                        if let Some(sb) = &mut self.sidebar {
-                            sb.toggle_selected();
-                        }
-                    } else {
-                        self.push_current_to_jump_list();
-                        let _ = self.editor.open_tab(path);
-                        self.after_file_open_or_save();
-                        self.sidebar_focused = false;
-                    }
-                }
-                true
-            }
-            EditorAction::InsertChar(' ') | EditorAction::MoveCursor(Direction::Right) => {
-                // Space / Right: open file (stay in sidebar) or expand/collapse directory.
-                let entry = self
-                    .sidebar
-                    .as_ref()
-                    .and_then(|sb| sb.entries.get(sb.selected))
-                    .map(|e| (e.path.clone(), e.is_dir));
-                if let Some((path, is_dir)) = entry {
-                    if is_dir {
-                        if let Some(sb) = &mut self.sidebar {
-                            sb.toggle_selected();
-                        }
-                    } else {
-                        let _ = self.editor.open_tab(path);
-                        self.after_file_open_or_save();
-                        // intentionally keep sidebar_focused = true
-                    }
-                }
-                true
-            }
-            EditorAction::MoveCursor(Direction::Left) => {
-                // Left: move to parent directory and collapse it.
-                if let Some(sb) = &mut self.sidebar {
-                    sb.move_to_parent_and_collapse();
-                }
-                self.ensure_sidebar_selected_visible();
-                true
-            }
-            EditorAction::FocusSidebar => {
-                // Ctrl+B while sidebar focused: jump back to editor, sidebar stays open.
-                self.sidebar_focused = false;
-                true
-            }
-            EditorAction::CopyFileReference => {
-                // Copy just the file path (no cursor location) when in sidebar.
-                let selected_path = self
-                    .sidebar
-                    .as_ref()
-                    .and_then(|sb| sb.selected_path().cloned());
-                if let Some(path) = selected_path {
-                    let reference = path
-                        .strip_prefix(&self.workspace)
-                        .unwrap_or(&path)
-                        .display()
-                        .to_string();
-                    self.clipboard.set(reference);
-                }
-                true
-            }
-            EditorAction::CloseSearch => {
-                // Esc: return focus to the editor without closing the sidebar.
-                self.sidebar_focused = false;
-                true
-            }
-            EditorAction::Copy => {
-                // Ctrl+C: copy file path to sidebar clipboard (not root).
-                let sel = self.sidebar.as_ref();
-                if sel.map(|sb| !sb.root_is_selected()).unwrap_or(false)
-                    && let Some(path) = sel.and_then(|sb| sb.selected_path().cloned())
-                {
-                    self.sidebar_clipboard = Some(SidebarClipboard {
-                        path,
-                        is_cut: false,
-                    });
-                }
-                true
-            }
-            EditorAction::Cut => {
-                // Ctrl+X: cut file path to sidebar clipboard (not root).
-                let sel = self.sidebar.as_ref();
-                if sel.map(|sb| !sb.root_is_selected()).unwrap_or(false)
-                    && let Some(path) = sel.and_then(|sb| sb.selected_path().cloned())
-                {
-                    self.sidebar_clipboard = Some(SidebarClipboard { path, is_cut: true });
-                }
-                true
-            }
-            EditorAction::Paste(_) => {
-                // Ctrl+V: paste (move or copy) the file from sidebar clipboard.
-                self.sidebar_paste();
-                true
-            }
-            EditorAction::DeleteForward => {
-                // Delete key: delete the selected file/directory (not root).
-                let is_root = self
-                    .sidebar
-                    .as_ref()
-                    .map(|sb| sb.root_is_selected())
-                    .unwrap_or(true);
-                if !is_root
-                    && let Some(path) = self
-                        .sidebar
-                        .as_ref()
-                        .and_then(|sb| sb.selected_path().cloned())
-                {
-                    if path.is_dir() {
-                        self.confirm_delete = Some(ConfirmDelete::Dir(path));
-                    } else {
-                        self.confirm_delete = Some(ConfirmDelete::File(path));
-                    }
-                }
-                true
-            }
-            EditorAction::RenameSymbol | EditorAction::SidebarRename => {
-                // F2: rename the selected file/directory (not root).
-                let is_root = self
-                    .sidebar
-                    .as_ref()
-                    .map(|sb| sb.root_is_selected())
-                    .unwrap_or(true);
-                if !is_root
-                    && let Some(path) = self
-                        .sidebar
-                        .as_ref()
-                        .and_then(|sb| sb.selected_path().cloned())
-                {
-                    let name = path
-                        .file_name()
-                        .and_then(|n| n.to_str())
-                        .unwrap_or("")
-                        .to_string();
-                    self.input_mode = InputMode::RenamePath(path, name);
-                }
-                true
-            }
-            EditorAction::SidebarNewFolder => {
-                // Ctrl+Shift+N: create a new folder in the selected location.
-                let parent = self.sidebar.as_ref().and_then(|sb| {
-                    sb.entries.get(sb.selected).map(|e| {
-                        if e.is_dir {
-                            e.path.clone()
-                        } else {
-                            e.path.parent().unwrap_or(&sb.root).to_path_buf()
-                        }
-                    })
-                });
-                if let Some(parent) = parent {
-                    self.input_mode = InputMode::NewFolderName(parent, String::new());
-                }
-                true
-            }
-            EditorAction::SidebarRefresh => {
-                self.refresh_sidebar();
-                true
-            }
-            // Global actions that don't touch editor content are allowed to
-            // fall through to the main dispatcher.
-            EditorAction::Quit
-            | EditorAction::ForceQuit
-            | EditorAction::ToggleHelp
-            | EditorAction::ToggleSidebar
-            | EditorAction::OpenSettings
-            | EditorAction::OpenCommandPalette
-            | EditorAction::OpenFuzzyPicker
-            | EditorAction::OpenRecentFiles
-            | EditorAction::OpenLspConfig
-            | EditorAction::OpenGitDialog
-            | EditorAction::ReloadConfig
-            | EditorAction::ToggleWordWrap
-            | EditorAction::OpenFile
-            | EditorAction::SaveFile
-            | EditorAction::SaveFileAs
-            | EditorAction::NewFile
-            | EditorAction::NewTab
-            | EditorAction::CloseTab
-            | EditorAction::NextTab
-            | EditorAction::PrevTab
-            | EditorAction::GoToTab(_)
-            | EditorAction::MouseClick { .. }
-            | EditorAction::MouseDrag { .. }
-            | EditorAction::MouseUp { .. }
-            | EditorAction::MouseScroll { .. }
-            | EditorAction::Unhandled => false,
-            // Swallow everything else so editor content / cursor / search /
-            // LSP state isn't affected while the sidebar has focus.
-            _ => true,
-        }
-    }
-
-    /// Paste from the sidebar clipboard into the currently selected location.
-    fn sidebar_paste(&mut self) {
-        let clip = match &self.sidebar_clipboard {
-            Some(c) => c,
-            None => return,
-        };
-        let dest_dir = match self.sidebar.as_ref() {
-            Some(sb) => match sb.entries.get(sb.selected) {
-                Some(entry) if entry.is_dir => entry.path.clone(),
-                Some(entry) => entry.path.parent().unwrap_or(&sb.root).to_path_buf(),
-                None => return,
-            },
-            None => return,
-        };
-        if clip.is_cut {
-            // Move: rename source into dest directory with collision check.
-            let source = clip.path.clone();
-            if let Some(name) = source.file_name() {
-                let new_path = dest_dir.join(name);
-                if new_path.exists() {
-                    return; // Don't overwrite existing files.
-                }
-                if std::fs::rename(&source, &new_path).is_ok() {
-                    // Only consume clipboard on success.
-                    self.sidebar_clipboard = None;
-                }
-            }
-        } else {
-            // Copy: only files (not directories).
-            let source = clip.path.clone();
-            if source.is_file()
-                && let Some(new_path) = copy_target_path(&source, &dest_dir)
-            {
-                let _ = std::fs::copy(&source, &new_path);
-            }
-            // Clipboard is kept so user can paste again.
-        }
-        self.refresh_sidebar();
-    }
-
-    /// Refresh the sidebar entries after a file operation.
-    fn refresh_sidebar(&mut self) {
-        if let Some(sb) = &mut self.sidebar {
-            sb.refresh();
-        }
-    }
-
-    /// Handle input while a delete confirmation is active.
-    fn handle_confirm_delete(&mut self, action: EditorAction) {
-        let state = self.confirm_delete.take();
-        match state {
-            Some(ConfirmDelete::File(path)) => match action {
-                EditorAction::InsertChar('y') | EditorAction::InsertChar('Y') => {
-                    let _ = std::fs::remove_file(&path);
-                    self.refresh_sidebar();
-                }
-                _ => {} // Any other key cancels.
-            },
-            Some(ConfirmDelete::Dir(path)) => match action {
-                EditorAction::InsertChar('y') | EditorAction::InsertChar('Y') => {
-                    // Move to second confirmation step.
-                    self.confirm_delete = Some(ConfirmDelete::DirConfirmed(path));
-                }
-                _ => {} // Any other key cancels.
-            },
-            Some(ConfirmDelete::DirConfirmed(path)) if action == EditorAction::InsertNewline => {
-                let _ = std::fs::remove_dir_all(&path);
-                self.refresh_sidebar();
-            }
-            Some(ConfirmDelete::DirConfirmed(_)) | None => {}
-        }
-    }
-
     // ── Command palette input handling ───────────────────────────────────────
-
-    fn handle_command_palette(&mut self, action: EditorAction) {
-        match action {
-            EditorAction::InsertChar(c) => {
-                if let Some(p) = &mut self.command_palette {
-                    let mut q = p.query.clone();
-                    q.push(c);
-                    p.update_query(q);
-                }
-            }
-            EditorAction::DeleteBackward => {
-                if let Some(p) = &mut self.command_palette {
-                    let mut q = p.query.clone();
-                    q.pop();
-                    p.update_query(q);
-                }
-            }
-            EditorAction::MoveCursor(Direction::Up) => {
-                if let Some(p) = &mut self.command_palette {
-                    p.move_up();
-                }
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                if let Some(p) = &mut self.command_palette {
-                    p.move_down();
-                }
-            }
-            EditorAction::InsertNewline => {
-                // Execute the selected command.
-                let dispatched = self
-                    .command_palette
-                    .as_ref()
-                    .and_then(|p| p.execute_selected());
-                self.command_palette = None;
-                if let Some(action) = dispatched {
-                    // Guard: don't re-open palette from itself.
-                    if !matches!(action, EditorAction::OpenCommandPalette) {
-                        let th = self.term_height;
-                        self.update(action, th);
-                    }
-                }
-            }
-            EditorAction::Quit | EditorAction::Unhandled | EditorAction::CloseSearch => {
-                self.command_palette = None;
-            }
-            _ => {}
-        }
-    }
 
     // ── Code formatting ───────────────────────────────────────────────────────
 
-    /// Short status-bar label showing the active indent style, e.g.
-    /// `"spaces:4"` or `"tabs:8"`.
-    pub fn indent_label(&self) -> String {
-        let (indent, _) = self.indent_for_active();
-        match indent.style {
-            crate::formatting::IndentStyle::Tabs => format!("tabs:{}", indent.width),
-            crate::formatting::IndentStyle::Spaces => format!("spaces:{}", indent.width),
-        }
-    }
-
-    /// Resolve the live indent rules for the active buffer's language,
-    /// merging project + global config and falling back to the legacy
-    /// `tab_size` and built-in defaults.
-    ///
-    /// Per-buffer `.editorconfig` overrides win over every config layer.
-    fn indent_for_active(
-        &self,
-    ) -> (
-        crate::formatting::IndentConfig,
-        crate::formatting::IndentRules,
-    ) {
-        let active = self.editor.active();
-        let lang = active.syntax.language;
-        let resolver = crate::formatting::FormattingResolver {
-            global: &self.config.formatting,
-            project: self.project_fmt.as_ref(),
-            legacy_tab_size: self.config.tab_size,
-        };
-        let mut indent = resolver.indent(lang);
-        if let Some(style) = active.editorconfig.indent_style {
-            indent.style = style;
-        }
-        if let Some(width) = active.editorconfig.effective_width()
-            && width > 0
-        {
-            indent.width = width;
-        }
-        (indent, crate::formatting::IndentRules::for_lang(lang))
-    }
-
-    /// Run the configured external formatter for the active buffer's
-    /// language and replace the buffer atomically. On any error, the buffer
-    /// is left untouched and `status_error` is set.
-    fn format_buffer(&mut self) {
-        let lang = self.editor.active().syntax.language;
-        let path = self.editor.active().path.clone();
-        let resolver = crate::formatting::FormattingResolver {
-            global: &self.config.formatting,
-            project: self.project_fmt.as_ref(),
-            legacy_tab_size: self.config.tab_size,
-        };
-        let fc = match resolver.formatter(lang) {
-            Some(f) => f,
-            None => {
-                let name = lang.name();
-                let display = if name.is_empty() { "this file" } else { name };
-                self.status_error = Some(format!("No formatter configured for {display}"));
-                return;
-            }
-        };
-
-        let input = self.editor.active().buffer.to_string();
-        let (saved_line, saved_col) = {
-            let c = self.editor.active().buffer.cursors.primary();
-            (c.line, c.col)
-        };
-
-        match crate::formatting::run_formatter(&fc, &input, path.as_deref()) {
-            Ok(out) if out == input => {
-                // No-op format — leave the buffer alone, no undo entry.
-            }
-            Ok(out) => {
-                let buf = &mut self.editor.active_mut().buffer;
-                buf.begin_batch();
-                let len = buf.rope().len_bytes();
-                buf.delete_range(0, len);
-                buf.move_cursor_to(0, false);
-                buf.insert_str(&out);
-                buf.commit_batch();
-                // Restore cursor by clamped (line, col).
-                let new_cursor =
-                    crate::buffer::cursor::Cursor::from_line_col(buf.rope(), saved_line, saved_col);
-                buf.move_cursor_to(new_cursor.byte_offset, false);
-                self.editor.active_mut().reparse();
-            }
-            Err(e) => {
-                self.status_error = Some(format!("Formatter failed: {e}"));
-            }
-        }
-    }
-
     // ── Line comment toggle ───────────────────────────────────────────────────
 
-    fn toggle_line_comment(&mut self) {
-        let prefix = match self.editor.active().syntax.comment_prefix() {
-            Some(p) => p,
-            None => return, // language has no line comment syntax
-        };
-        let cursor_line = self.editor.active().buffer.cursors.primary().line;
-        let line_str = self.editor.active().buffer.line_str(cursor_line);
-        let trimmed = line_str.trim_start();
-        let leading_spaces = line_str.len() - trimmed.len();
-        let already_commented = trimmed.starts_with(prefix);
-
-        let buf = &mut self.editor.active_mut().buffer;
-        let line_start = buf
-            .rope()
-            .char_to_byte(buf.rope().line_to_char(cursor_line));
-
-        buf.begin_batch();
-        if already_commented {
-            // Remove the comment prefix.
-            let comment_start = line_start + leading_spaces;
-            let comment_end = comment_start + prefix.len();
-            buf.delete_range(comment_start, comment_end);
-        } else {
-            // Insert the comment prefix at the start of the line.
-            buf.move_cursor_to(line_start, false);
-            buf.insert_str(prefix);
-        }
-        buf.commit_batch();
-    }
-
     // ── File helpers ─────────────────────────────────────────────────────────
-
-    fn close_tab(&mut self) {
-        if self.editor.active().buffer.modified && self.config.confirm_exit {
-            self.confirm_quit = true; // reuse confirm for "discard changes?"
-        } else {
-            // Notify LSP before closing.
-            if let Some(path) = self.editor.active().path.clone() {
-                self.notify_lsp_did_close(&path);
-            }
-            self.editor.close_active_tab();
-        }
-    }
-
-    fn save_active(&mut self) {
-        if self.editor.active().path.is_some() {
-            let _ = self.editor.active_mut().save();
-            self.after_file_open_or_save();
-            // Flush any pending didChange before sending didSave.
-            if self.lsp_dirty_since.is_some() && !self.lsp_change_sent {
-                self.send_lsp_did_change();
-            }
-            self.lsp_dirty_since = None;
-            self.lsp_change_sent = false;
-            self.notify_lsp_did_save();
-            // Persist undo history (opt-in) so Ctrl+Z still works after a
-            // restart. Silent on I/O failure.
-            if self.config.persistent_undo {
-                let tab = self.editor.active();
-                if let Some(path) = &tab.path {
-                    let content = tab.buffer.to_string();
-                    let snap = tab.buffer.history_snapshot();
-                    crate::buffer::persistent_undo::save(&self.workspace, path, &content, &snap);
-                }
-            }
-        } else {
-            self.input_mode = InputMode::SaveAsPath(String::new());
-        }
-    }
-
-    /// Owned copy of the row-0 gutter badge string ("↑X.Y.Z") that
-    /// `editor_view::render` overlays when a newer release is available. Used
-    /// by the layout helpers to widen the gutter width consistently across
-    /// renderer and mouse hit-testing.
-    pub fn version_badge(&self) -> Option<String> {
-        self.version_check.newer_version().map(|v| format!("↑{v}"))
-    }
-
-    /// Recompute the git gutter for the currently active buffer (if it has a path).
-    /// Move the cursor to the start of the next (`step == 1`) or previous
-    /// (`step == -1`) git hunk. No-op when there are no hunks or git is off.
-    fn jump_to_relative_hunk(&mut self, step: i32) {
-        let Some(gutter) = self.git_gutter.as_ref() else {
-            return;
-        };
-        let hunks = gutter.hunks();
-        if hunks.is_empty() {
-            self.status_error = Some("No git hunks in this buffer".into());
-            return;
-        }
-        let cur_line = self.editor.active().buffer.cursors.primary().line;
-        let target = if step > 0 {
-            hunks
-                .iter()
-                .find(|h| h.start_line > cur_line)
-                .copied()
-                .unwrap_or(hunks[0])
-        } else {
-            hunks
-                .iter()
-                .rev()
-                .find(|h| h.end_line < cur_line)
-                .copied()
-                .unwrap_or(*hunks.last().unwrap())
-        };
-        self.push_current_to_jump_list();
-        let rope = self.editor.active().buffer.rope().clone();
-        let line = target.start_line.min(rope.len_lines().saturating_sub(1));
-        let cursor = crate::buffer::cursor::Cursor::from_line_col(&rope, line, 0);
-        *self.editor.active_mut().buffer.cursors.primary_mut() = cursor;
-    }
-
-    /// Replace the hunk containing the cursor with its HEAD content.
-    fn revert_hunk_at_cursor(&mut self) {
-        let Some(gutter) = self.git_gutter.as_ref() else {
-            self.status_error = Some("No git gutter available".into());
-            return;
-        };
-        let cur_line = self.editor.active().buffer.cursors.primary().line;
-        let hunk_opt = gutter
-            .hunks()
-            .into_iter()
-            .find(|h| cur_line >= h.start_line && cur_line <= h.end_line);
-        let Some(hunk) = hunk_opt else {
-            self.status_error = Some("Cursor is not inside a hunk".into());
-            return;
-        };
-        let Some(path) = self.editor.active().path.clone() else {
-            self.status_error = Some("Save the file before reverting a hunk".into());
-            return;
-        };
-        let head_content = match crate::git::fetch_head_content(&path) {
-            Some(c) => c,
-            None => {
-                self.status_error = Some("File is not tracked in HEAD".into());
-                return;
-            }
-        };
-        // Recompute the hunk's HEAD-side line span by re-running the same
-        // diff that produced the gutter. We need to find the HEAD lines that
-        // map to this hunk's [start_line, end_line] range in the current
-        // buffer.
-        let current_full = self.editor.active().buffer.to_string();
-        let head_lines: Vec<&str> = head_content.lines().collect();
-        let current_lines: Vec<&str> = current_full.lines().collect();
-        let (head_start, head_end) = crate::git::head_range_for_hunk(
-            &head_lines,
-            &current_lines,
-            hunk.start_line,
-            hunk.end_line,
-        );
-
-        // Replace the buffer slice [hunk.start_line, hunk.end_line] with the
-        // HEAD content slice [head_start, head_end].
-        let replacement = if head_start <= head_end && head_end < head_lines.len() {
-            let mut s = head_lines[head_start..=head_end].join("\n");
-            // Keep the trailing newline so we don't merge the next line into
-            // the hunk's last line.
-            s.push('\n');
-            s
-        } else {
-            // Pure deletion in HEAD — replacement is empty (matches: hunk
-            // exists in current but has no HEAD counterpart).
-            String::new()
-        };
-
-        let buf = &mut self.editor.active_mut().buffer;
-        let start_byte = buf.line_start_byte(hunk.start_line);
-        let end_line = (hunk.end_line + 1).min(buf.len_lines());
-        let end_byte = buf.line_start_byte(end_line);
-        buf.begin_batch();
-        buf.delete_range(start_byte, end_byte);
-        if !replacement.is_empty() {
-            buf.cursors.primary_mut().byte_offset = start_byte;
-            buf.insert_str(&replacement);
-        }
-        buf.commit_batch();
-        self.refresh_git_gutter();
-    }
-
-    /// Toggle the inline diff-peek float for the hunk under the cursor.
-    fn toggle_diff_peek(&mut self) {
-        if self.diff_peek.is_some() {
-            self.diff_peek = None;
-            return;
-        }
-        let Some(gutter) = self.git_gutter.as_ref() else {
-            return;
-        };
-        let cur_line = self.editor.active().buffer.cursors.primary().line;
-        let Some(hunk) = gutter
-            .hunks()
-            .into_iter()
-            .find(|h| cur_line >= h.start_line && cur_line <= h.end_line)
-        else {
-            self.status_error = Some("Cursor is not inside a hunk".into());
-            return;
-        };
-        let Some(path) = self.editor.active().path.clone() else {
-            return;
-        };
-        let Some(head_content) = crate::git::fetch_head_content(&path) else {
-            self.status_error = Some("File is not tracked in HEAD".into());
-            return;
-        };
-        let current_full = self.editor.active().buffer.to_string();
-        let head_lines: Vec<&str> = head_content.lines().collect();
-        let current_lines: Vec<&str> = current_full.lines().collect();
-        let (head_start, head_end) = crate::git::head_range_for_hunk(
-            &head_lines,
-            &current_lines,
-            hunk.start_line,
-            hunk.end_line,
-        );
-        let lines: Vec<String> = if head_start <= head_end && head_end < head_lines.len() {
-            head_lines[head_start..=head_end]
-                .iter()
-                .map(|s| s.to_string())
-                .collect()
-        } else {
-            vec!["(no HEAD content for this hunk)".into()]
-        };
-        self.diff_peek = Some(DiffPeekState {
-            head_lines: lines,
-            anchor_line: cur_line,
-        });
-    }
-
-    /// Build a `Session` snapshot of the current editor state and write it
-    /// to `<workspace>/.txt/session.json`. Called once on clean shutdown
-    /// when `restore_session = true`. Buffers without a saved path are
-    /// skipped — they have nothing to reopen on the next launch.
-    fn save_session(&self) {
-        use crate::session::{Session, TabState};
-        let mut tabs = Vec::new();
-        let mut active_in_session = 0usize;
-        let active_idx = self.editor.active_idx;
-        for (i, tab) in self.editor.tabs.iter().enumerate() {
-            let Some(path) = &tab.path else {
-                continue;
-            };
-            if i <= active_idx {
-                active_in_session = tabs.len();
-            }
-            tabs.push(TabState {
-                path: path.clone(),
-                cursor_byte: tab.buffer.cursors.primary().byte_offset,
-                viewport_top: tab.viewport.scroll_row,
-            });
-        }
-        let session = Session {
-            tabs,
-            active: active_in_session,
-            sidebar_open: self.sidebar.is_some(),
-        };
-        session.save(&self.workspace);
-    }
-
-    /// Open every tab listed in `session` and restore cursor positions /
-    /// viewport tops. Files that no longer exist on disk are skipped.
-    /// Returns `true` when at least one tab was restored.
-    pub fn restore_from_session(&mut self, session: crate::session::Session) -> bool {
-        let mut restored_any = false;
-        for tab in &session.tabs {
-            if !tab.path.exists() {
-                continue;
-            }
-            if self.editor.open_tab(tab.path.clone()).is_ok() {
-                let buf = &mut self.editor.active_mut().buffer;
-                let bound = tab.cursor_byte.min(buf.rope().len_bytes());
-                *buf.cursors.primary_mut() =
-                    crate::buffer::cursor::Cursor::from_byte_offset(buf.rope(), bound);
-                self.editor.active_mut().viewport.scroll_row = tab.viewport_top;
-                restored_any = true;
-            }
-        }
-        if restored_any && session.active < self.editor.tabs.len() {
-            self.editor.active_idx = session.active;
-        }
-        if session.sidebar_open && self.sidebar.is_none() {
-            self.sidebar = Some(SidebarState::new());
-        }
-        restored_any
-    }
-
-    fn refresh_git_gutter(&mut self) {
-        let path = self.editor.active().path.clone();
-        if let Some(path) = path {
-            let content = self.editor.active().buffer.to_string();
-            self.git_gutter = crate::git::gutter_for_path(&path, &content);
-        } else {
-            self.git_gutter = None;
-        }
-    }
-
-    /// Replay the actions stored in macro slot `slot`. Wraps the playback in
-    /// a single undo batch so the entire sequence collapses to one undo step,
-    /// and sets the replay flag so the actions are not re-recorded if the
-    /// user starts a new recording mid-replay.
-    fn replay_macro_slot(&mut self, slot: char) {
-        let Some(actions) = self.macros.play(slot) else {
-            self.status_error = Some(format!("No macro in slot '{slot}'"));
-            return;
-        };
-        let term_h = self.term_height;
-        self.macros.set_replaying(true);
-        self.editor.active_mut().buffer.begin_batch();
-        for action in actions {
-            self.update(action, term_h);
-        }
-        self.editor.active_mut().buffer.commit_batch();
-        self.macros.set_replaying(false);
-    }
-
-    /// Same logic as `expand_snippet_at_cursor` but silent: returns `false`
-    /// when no snippet matches so the caller can fall through to its own
-    /// default (e.g. `InsertTab` doing indentation).
-    fn try_expand_snippet_silently(&mut self) -> bool {
-        let lang_id = self.editor.active().syntax.language.config_key();
-        if lang_id.is_empty() {
-            return false;
-        }
-        let cursor_byte = self.editor.active().buffer.cursors.primary().byte_offset;
-        let primary = self.editor.active().buffer.cursors.primary();
-        if primary.has_selection() {
-            return false;
-        }
-        let rope = self.editor.active().buffer.rope();
-        let probe = cursor_byte.saturating_sub(1);
-        let Some((wstart, wend)) = crate::buffer::cursor::word_span_at(rope, probe) else {
-            return false;
-        };
-        if wend < cursor_byte {
-            return false;
-        }
-        let prefix: String = rope.byte_slice(wstart..wend).chars().collect();
-        let matches = self.snippets.lookup(lang_id, &prefix);
-        let Some(snip) = matches.into_iter().next() else {
-            return false;
-        };
-        let parsed = snip.parse_body();
-        let exp = crate::snippet::session::SnippetSession::expand_at(&parsed, wstart);
-        let buf = &mut self.editor.active_mut().buffer;
-        buf.begin_batch();
-        buf.delete_range(wstart, wend);
-        *buf.cursors.primary_mut() =
-            crate::buffer::cursor::Cursor::from_byte_offset(buf.rope(), wstart);
-        buf.cursors.primary_mut().selection = None;
-        buf.insert_str(&exp.text);
-        buf.commit_batch();
-        self.editor.active_mut().snippet_session = exp.session;
-        self.snippet_select_current();
-        true
-    }
-
-    /// Look up the word before the cursor in the active buffer's language
-    /// snippet store; if there's a match, delete the word and expand the
-    /// snippet in its place.
-    fn expand_snippet_at_cursor(&mut self) {
-        let lang_id = self.editor.active().syntax.language.config_key();
-        if lang_id.is_empty() {
-            self.status_error = Some("Snippets need a recognised language".into());
-            return;
-        }
-        let cursor_byte = self.editor.active().buffer.cursors.primary().byte_offset;
-        let rope = self.editor.active().buffer.rope();
-        let probe = cursor_byte.saturating_sub(1);
-        let Some((wstart, wend)) = crate::buffer::cursor::word_span_at(rope, probe) else {
-            self.status_error = Some("Place the cursor after a snippet prefix".into());
-            return;
-        };
-        if wend < cursor_byte {
-            self.status_error = Some("Place the cursor after a snippet prefix".into());
-            return;
-        }
-        let prefix: String = rope.byte_slice(wstart..wend).chars().collect();
-        let matches = self.snippets.lookup(lang_id, &prefix);
-        let Some(snip) = matches.into_iter().next() else {
-            self.status_error = Some(format!("No snippet named '{prefix}'"));
-            return;
-        };
-        let parsed = snip.parse_body();
-        // Compute the expansion text and session at the prefix start position
-        // since we're about to delete the prefix.
-        let exp = crate::snippet::session::SnippetSession::expand_at(&parsed, wstart);
-        let buf = &mut self.editor.active_mut().buffer;
-        buf.begin_batch();
-        buf.delete_range(wstart, wend);
-        // Move the primary cursor to wstart before inserting so insert lands
-        // at the right location regardless of prior selection state.
-        *buf.cursors.primary_mut() =
-            crate::buffer::cursor::Cursor::from_byte_offset(buf.rope(), wstart);
-        buf.cursors.primary_mut().selection = None;
-        buf.insert_str(&exp.text);
-        buf.commit_batch();
-        self.editor.active_mut().snippet_session = exp.session;
-        // Jump cursor to the first tab stop, selecting its default text.
-        self.snippet_select_current();
-    }
-
-    /// Select the byte range of the current snippet tab stop. Called after
-    /// expanding or advancing the session.
-    fn snippet_select_current(&mut self) {
-        let handle = self.editor.active_mut();
-        let range = match &handle.snippet_session {
-            Some(s) => s.current_range(),
-            None => return,
-        };
-        let len = handle.buffer.rope().len_bytes();
-        let bound_start = range.start.min(len);
-        let bound_end = range.end.min(len);
-        let new_cursor = {
-            let rope = handle.buffer.rope();
-            crate::buffer::cursor::Cursor::from_byte_offset(rope, bound_end)
-        };
-        let primary = handle.buffer.cursors.primary_mut();
-        *primary = new_cursor;
-        if bound_end > bound_start {
-            primary.selection = Some(crate::buffer::cursor::Selection {
-                anchor: bound_start,
-                active: bound_end,
-            });
-        } else {
-            primary.selection = None;
-        }
-    }
-
-    /// Move the active snippet session forward (`true`) or backward.
-    fn snippet_advance(&mut self, forward: bool) {
-        let advanced = match self.editor.active_mut().snippet_session.as_mut() {
-            Some(s) => {
-                if forward {
-                    s.next_stop()
-                } else {
-                    s.prev_stop()
-                }
-            }
-            None => false,
-        };
-        if advanced {
-            self.snippet_select_current();
-        } else {
-            self.editor.active_mut().snippet_session = None;
-        }
-    }
-
-    /// Snapshot the active buffer's cursor into the jump list. Used right
-    /// before navigation actions (mark jumps, jump-list back) so the
-    /// previous cursor position can be returned to.
-    fn push_current_to_jump_list(&mut self) {
-        let handle = self.editor.active();
-        if let Some(path) = handle.path.clone() {
-            let byte_offset = handle.buffer.cursors.primary().byte_offset;
-            self.jumps
-                .push(crate::marks::JumpEntry { path, byte_offset });
-        }
-    }
-
-    /// Move the cursor to a jump-list entry, switching tabs if necessary.
-    fn go_to_jump_entry(&mut self, entry: &crate::marks::JumpEntry) {
-        let active_path = self.editor.active().path.clone();
-        if active_path.as_ref() != Some(&entry.path) {
-            let _ = self.editor.open_tab(entry.path.clone());
-            self.after_file_open_or_save();
-        }
-        let handle = self.editor.active_mut();
-        let rope = handle.buffer.rope();
-        let bound = entry.byte_offset.min(rope.len_bytes());
-        *handle.buffer.cursors.primary_mut() =
-            crate::buffer::cursor::Cursor::from_byte_offset(rope, bound);
-        handle.buffer.cursors.collapse_to_primary();
-    }
-
-    /// Called after a file is opened or saved — updates recent files, git gutter,
-    /// installs a file watcher, and notifies the LSP server.
-    /// Attempt to load persistent undo history for the active tab. No-op
-    /// when the feature is disabled, the buffer has no path, or the
-    /// on-disk hash doesn't match the buffer content.
-    pub fn try_load_persistent_undo_for_active(&mut self) {
-        if !self.config.persistent_undo {
-            return;
-        }
-        let path = match self.editor.active().path.clone() {
-            Some(p) => p,
-            None => return,
-        };
-        let content = self.editor.active().buffer.to_string();
-        if let Some(snap) = crate::buffer::persistent_undo::load(&self.workspace, &path, &content) {
-            self.editor.active_mut().buffer.restore_history(snap);
-        }
-    }
-
-    fn after_file_open_or_save(&mut self) {
-        if let Some(path) = self.editor.active().path.clone() {
-            add_to_recent_files(&path, &self.workspace.clone());
-            self.file_watcher = FileWatcher::new(&path);
-        }
-        // Persistent undo: a freshly opened buffer has an empty history, so
-        // we use that as the cue that this is an `open` event (not a
-        // `save`). This avoids the load clobbering the in-memory history
-        // that we are about to persist on save.
-        if !self.editor.active().buffer.can_undo() {
-            self.try_load_persistent_undo_for_active();
-        }
-        self.refresh_git_gutter();
-        // Notify LSP server that a file was opened.
-        let handle = self.editor.active();
-        // Avoid borrow conflict by extracting what we need.
-        let path = handle.path.clone();
-        let lang = handle.syntax.language.name().to_lowercase();
-        let version = handle.lsp_state.version;
-        let text = handle.buffer.rope().to_string();
-        if let Some(registry) = &self.lsp
-            && registry.is_ready()
-            && let Some(path) = &path
-        {
-            let uri = crate::lsp::types::path_to_uri(path);
-            let _ = registry.client().did_open(&uri, &lang, version, &text);
-        }
-    }
-
-    /// Poll the file watcher; if the file changed externally, reload automatically.
-    pub fn poll_file_watcher(&mut self) {
-        if let Some(watcher) = &self.file_watcher
-            && watcher.poll()
-        {
-            self.reload_active_file();
-        }
-    }
-
-    /// Save the active buffer automatically after 1 second of inactivity (debounced).
-    ///
-    /// Only saves when `config.auto_save` is enabled and the buffer has a path.
-    pub fn poll_auto_save(&mut self) {
-        if !self.config.auto_save {
-            return;
-        }
-        if let Some(t) = self.auto_save_timer
-            && t.elapsed() >= std::time::Duration::from_secs(1)
-            && self.editor.active().path.is_some()
-        {
-            self.save_active();
-            self.auto_save_timer = None;
-        }
-    }
-
-    /// Update cached RSS memory usage (throttled to every 2 seconds).
-    pub fn refresh_memory(&mut self) {
-        if self.memory_last_checked.elapsed() >= Duration::from_secs(2) {
-            if let Some(kb) = read_rss_kb() {
-                self.memory_rss_kb = kb;
-            }
-            self.memory_last_checked = Instant::now();
-        }
-    }
-
-    /// Refresh the cached git branch (throttled to every 2 seconds).
-    ///
-    /// Picks up branch changes made outside the editor (e.g. `git checkout`
-    /// from another terminal).
-    pub fn refresh_git_branch(&mut self) {
-        if self.git_branch_last_checked.elapsed() >= Duration::from_secs(2) {
-            self.git_branch = crate::git::current_branch(&self.workspace);
-            self.git_branch_last_checked = Instant::now();
-        }
-    }
-
-    /// Reload the active buffer from disk (used after external modification).
-    fn reload_active_file(&mut self) {
-        let path = match self.editor.active().path.clone() {
-            Some(p) => p,
-            None => return,
-        };
-        if let Ok(text) = std::fs::read_to_string(&path) {
-            let handle = self.editor.active_mut();
-            let saved_line = handle.buffer.cursors.primary().line;
-            let saved_col = handle.buffer.cursors.primary().col;
-            handle.buffer = crate::buffer::Buffer::from_str(&text);
-            handle.buffer.modified = false;
-            let rope = handle.buffer.rope().clone();
-            *handle.buffer.cursors.primary_mut() =
-                crate::buffer::cursor::Cursor::from_line_col(&rope, saved_line, saved_col);
-        }
-        self.refresh_git_gutter();
-        // Re-install watcher after reload so we don't miss the next change.
-        if let Some(path) = self.editor.active().path.clone() {
-            self.file_watcher = FileWatcher::new(&path);
-        }
-    }
 
     // ── LSP polling ──────────────────────────────────────────────────────────
 
@@ -3945,1364 +1473,23 @@ impl AppState {
     /// How long to wait after the last edit before re-requesting semantic tokens.
     const SEMANTIC_TOKEN_DEBOUNCE: Duration = Duration::from_millis(300);
 
-    /// Flush debounced LSP notifications if enough idle time has passed.
-    /// Called once per frame in the event loop.
-    pub fn flush_lsp_debounce(&mut self) {
-        let Some(dirty_since) = self.lsp_dirty_since else {
-            return;
-        };
-        let elapsed = dirty_since.elapsed();
-
-        // After 100ms idle, send the buffered didChange (one full-buffer copy).
-        if elapsed >= Self::LSP_DEBOUNCE && !self.lsp_change_sent {
-            self.send_lsp_did_change();
-            self.lsp_change_sent = true;
-        }
-
-        // After 300ms idle, re-request semantic tokens and clear the timer.
-        if elapsed >= Self::SEMANTIC_TOKEN_DEBOUNCE {
-            self.request_semantic_tokens_for_active();
-            self.lsp_dirty_since = None;
-            self.lsp_change_sent = false;
-        }
-    }
-
-    /// Non-blocking drain of pending LSP updates. Called once per frame.
-    pub fn poll_lsp_updates(&mut self) {
-        let Some(registry) = &mut self.lsp else {
-            return;
-        };
-        let updates = registry.poll();
-        for update in updates {
-            self.apply_lsp_update(update);
-        }
-    }
-
-    fn apply_lsp_update(&mut self, update: crate::lsp::client::LspUpdate) {
-        use crate::lsp::client::LspUpdate;
-        match update {
-            LspUpdate::Initialized(caps) => {
-                if let Some(registry) = &mut self.lsp {
-                    registry.client_mut().capabilities = caps;
-                    registry.client_mut().initialized = true;
-                    let _ = registry
-                        .client()
-                        .send_notification("initialized", Some(serde_json::json!({})));
-                }
-                // Send didOpen for all currently open buffers.
-                self.notify_lsp_did_open_all();
-                // Request semantic tokens for the active buffer.
-                self.request_semantic_tokens_for_active();
-            }
-            LspUpdate::Diagnostics { uri, diagnostics } => {
-                self.apply_diagnostics(&uri, &diagnostics);
-            }
-            LspUpdate::ServerExited => {
-                // If the binary on disk is still the one the user approved,
-                // use the existing in-place restart path (preserves restart_count
-                // so a crash loop self-terminates after MAX_RESTARTS).
-                // If the binary changed (or vanished), tear down and route
-                // through the approval gate so the user is re-prompted.
-                let trusted = self.lsp_binary_still_trusted();
-
-                if !trusted {
-                    self.lsp = None;
-                    self.status_error =
-                        Some("LSP binary changed since approval; re-prompting".into());
-                    self.request_lsp_start();
-                    return;
-                }
-
-                let config = self.lsp_config.clone();
-                let workspace = self.workspace.clone();
-                let mut disable_lsp = false;
-
-                if let Some(registry) = &mut self.lsp
-                    && (registry.restart_exhausted()
-                        || registry.try_restart(&config, &workspace).is_err())
-                {
-                    disable_lsp = true;
-                }
-
-                if disable_lsp {
-                    self.lsp = None;
-                    self.status_error =
-                        Some("LSP server exited unexpectedly (restart limit reached)".into());
-                } else {
-                    self.status_error = Some("LSP server exited, restarting…".into());
-                }
-            }
-            LspUpdate::Completion { items, .. } => {
-                self.apply_completion_response(items);
-            }
-            LspUpdate::Hover { contents, .. } => {
-                self.apply_hover_response(contents);
-            }
-            LspUpdate::Definition { locations, .. } => {
-                self.apply_definition_response(locations);
-            }
-            LspUpdate::References { locations, .. } => {
-                self.apply_references_response(locations);
-            }
-            LspUpdate::Rename { edit, .. } => {
-                if let Some(edit) = edit {
-                    self.apply_workspace_edit(&edit);
-                }
-            }
-            LspUpdate::CodeActions { actions, .. } => {
-                let _ = actions; // TODO: show code action picker
-            }
-            LspUpdate::SemanticTokens { uri, data } => {
-                self.apply_semantic_tokens(&uri, &data);
-            }
-            LspUpdate::Error(msg) => {
-                self.status_error = Some(msg);
-            }
-        }
-    }
-
-    fn notify_lsp_did_open_all(&self) {
-        let Some(registry) = &self.lsp else { return };
-        if !registry.is_ready() {
-            return;
-        }
-        for tab in &self.editor.tabs {
-            if let Some(path) = &tab.path {
-                let uri = crate::lsp::types::path_to_uri(path);
-                let lang_id = tab.syntax.language.name().to_lowercase();
-                let text = tab.buffer.rope().to_string();
-                let _ = registry
-                    .client()
-                    .did_open(&uri, &lang_id, tab.lsp_state.version, &text);
-            }
-        }
-    }
-
-    /// Send `textDocument/didOpen` for a single buffer.
-    #[allow(dead_code)]
-    fn notify_lsp_did_open(&self, handle: &crate::editor::tab::BufferHandle) {
-        let Some(registry) = &self.lsp else { return };
-        if !registry.is_ready() {
-            return;
-        }
-        if let Some(path) = &handle.path {
-            let uri = crate::lsp::types::path_to_uri(path);
-            let lang_id = handle.syntax.language.name().to_lowercase();
-            let text = handle.buffer.rope().to_string();
-            let _ = registry
-                .client()
-                .did_open(&uri, &lang_id, handle.lsp_state.version, &text);
-        }
-    }
-
-    /// Send `textDocument/didChange` for the active buffer (full sync).
-    /// Version must already be bumped before calling this.
-    fn send_lsp_did_change(&self) {
-        let Some(registry) = &self.lsp else { return };
-        if !registry.is_ready() {
-            return;
-        }
-        let handle = self.editor.active();
-        if let Some(path) = &handle.path {
-            let uri = crate::lsp::types::path_to_uri(path);
-            let version = handle.lsp_state.version;
-            let text = handle.buffer.rope().to_string();
-            let _ = registry.client().did_change(&uri, version, &text);
-        }
-    }
-
-    /// Send `textDocument/didSave` for the active buffer.
-    fn notify_lsp_did_save(&self) {
-        let Some(registry) = &self.lsp else { return };
-        if !registry.is_ready() {
-            return;
-        }
-        if let Some(path) = &self.editor.active().path {
-            let uri = crate::lsp::types::path_to_uri(path);
-            let _ = registry.client().did_save(&uri);
-        }
-    }
-
-    /// Send `textDocument/didClose` for a buffer by path.
-    fn notify_lsp_did_close(&self, path: &std::path::Path) {
-        let Some(registry) = &self.lsp else { return };
-        if !registry.is_ready() {
-            return;
-        }
-        let uri = crate::lsp::types::path_to_uri(path);
-        let _ = registry.client().did_close(&uri);
-    }
-
-    /// Convert raw diagnostic JSON from the server to byte-offset `LspDiagnostic`s
-    /// and store them on the matching buffer.
-    fn apply_diagnostics(&mut self, uri: &str, raw_diagnostics: &[serde_json::Value]) {
-        use crate::lsp::types::{DiagSeverity, LspDiagnostic, lsp_position_to_byte_offset};
-
-        let path = match crate::lsp::types::uri_to_path(uri) {
-            Some(p) => p,
-            None => return,
-        };
-
-        // Find the buffer that matches this URI.
-        let tab = self
-            .editor
-            .tabs
-            .iter_mut()
-            .find(|t| t.path.as_ref().is_some_and(|p| same_file(p, &path)));
-        let Some(tab) = tab else { return };
-
-        let rope = tab.buffer.rope();
-        let mut diagnostics = Vec::with_capacity(raw_diagnostics.len());
-
-        for raw in raw_diagnostics {
-            let range = match raw.get("range") {
-                Some(r) => r,
-                None => continue,
-            };
-            let start = match parse_lsp_position(range.get("start")) {
-                Some(pos) => match lsp_position_to_byte_offset(rope, pos) {
-                    Some(b) => b,
-                    None => continue,
-                },
-                None => continue,
-            };
-            let end = match parse_lsp_position(range.get("end")) {
-                Some(pos) => match lsp_position_to_byte_offset(rope, pos) {
-                    Some(b) => b,
-                    None => continue,
-                },
-                None => continue,
-            };
-            let severity = match raw.get("severity").and_then(|v| v.as_u64()) {
-                Some(1) => DiagSeverity::Error,
-                Some(2) => DiagSeverity::Warning,
-                Some(3) => DiagSeverity::Information,
-                _ => DiagSeverity::Hint,
-            };
-            let message = raw
-                .get("message")
-                .and_then(|v| v.as_str())
-                .unwrap_or("")
-                .to_string();
-            let source = raw.get("source").and_then(|v| v.as_str()).map(String::from);
-
-            diagnostics.push(LspDiagnostic {
-                range: crate::buffer::cursor::ByteRange { start, end },
-                severity,
-                message,
-                source,
-            });
-        }
-
-        tab.lsp_state.diagnostics = diagnostics;
-    }
-
     // ── Completion ───────────────────────────────────────────────────────────
-
-    fn trigger_completion(&mut self) {
-        let Some(registry) = &mut self.lsp else {
-            return;
-        };
-        if !registry.is_ready() {
-            return;
-        }
-        let handle = self.editor.active();
-        let Some(path) = &handle.path else { return };
-        let uri = crate::lsp::types::path_to_uri(path);
-        let cursor = handle.buffer.cursors.primary();
-        let pos = crate::lsp::types::byte_offset_to_lsp_position(
-            handle.buffer.rope(),
-            cursor.byte_offset,
-        );
-        let anchor_byte = cursor.byte_offset;
-        let anchor_line = cursor.line;
-        let anchor_col = cursor.col;
-
-        let _ = registry
-            .client_mut()
-            .request_completion(&uri, pos.line, pos.character);
-
-        self.completion = Some(CompletionState::new(anchor_byte, anchor_line, anchor_col));
-    }
-
-    fn handle_completion_input(&mut self, action: &EditorAction) -> bool {
-        match action {
-            EditorAction::MoveCursor(Direction::Up) => {
-                if let Some(c) = &mut self.completion {
-                    c.selected = c.selected.saturating_sub(1);
-                }
-                true
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                if let Some(c) = &mut self.completion
-                    && !c.filtered.is_empty()
-                {
-                    c.selected = (c.selected + 1).min(c.filtered.len() - 1);
-                }
-                true
-            }
-            EditorAction::InsertNewline | EditorAction::InsertTab => {
-                self.accept_completion();
-                true
-            }
-            EditorAction::CloseSearch => {
-                self.completion = None;
-                true
-            }
-            // Cursor movement dismisses completion.
-            EditorAction::MoveCursor(_)
-            | EditorAction::MoveCursorWord(_)
-            | EditorAction::MoveCursorHome
-            | EditorAction::MoveCursorEnd => {
-                self.completion = None;
-                false // let the movement fall through
-            }
-            // Characters fall through to editing, then refilter.
-            _ => false,
-        }
-    }
-
-    fn accept_completion(&mut self) {
-        let insert_text = match &self.completion {
-            Some(c) => match c.selected_item() {
-                Some(item) => item.insert_text.clone(),
-                None => {
-                    self.completion = None;
-                    return;
-                }
-            },
-            None => return,
-        };
-        let anchor = self.completion.as_ref().unwrap().anchor_byte;
-        let cursor_byte = self.editor.active().buffer.cursors.primary().byte_offset;
-
-        // Delete the typed prefix and insert the completion text.
-        if cursor_byte > anchor {
-            let rope = self.editor.active().buffer.rope();
-            let start_char = rope.byte_to_char(anchor);
-            let end_char = rope.byte_to_char(cursor_byte);
-            self.editor
-                .active_mut()
-                .buffer
-                .delete_range(start_char, end_char);
-        }
-        self.editor.active_mut().buffer.insert_str(&insert_text);
-
-        self.completion = None;
-    }
-
-    fn refilter_completion(&mut self) {
-        let Some(comp) = &mut self.completion else {
-            return;
-        };
-        let cursor_byte = self.editor.active().buffer.cursors.primary().byte_offset;
-        if cursor_byte < comp.anchor_byte {
-            self.completion = None;
-            return;
-        }
-        let rope = self.editor.active().buffer.rope();
-        let start = rope.byte_to_char(comp.anchor_byte);
-        let end = rope.byte_to_char(cursor_byte);
-        let prefix: String = rope.slice(start..end).chars().collect();
-        comp.filter(&prefix);
-    }
-
-    fn apply_completion_response(&mut self, items: Vec<serde_json::Value>) {
-        let Some(comp) = &mut self.completion else {
-            return;
-        };
-        comp.items = items
-            .iter()
-            .map(|item| {
-                let label = item
-                    .get("label")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
-                let detail = item
-                    .get("detail")
-                    .and_then(|v| v.as_str())
-                    .map(String::from);
-                let insert_text = item
-                    .get("insertText")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or(&label)
-                    .to_string();
-                let filter_text = item
-                    .get("filterText")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or(&label)
-                    .to_string();
-                let kind = item.get("kind").and_then(|v| v.as_u64()).unwrap_or(0);
-                let kind_label = completion_kind_label(kind);
-                CompletionItemEntry {
-                    label,
-                    detail,
-                    insert_text,
-                    filter_text,
-                    kind_label,
-                }
-            })
-            .collect();
-        comp.filtered = (0..comp.items.len()).collect();
-        self.refilter_completion();
-    }
 
     // ── Hover ────────────────────────────────────────────────────────────────
 
-    fn trigger_hover(&mut self) {
-        let Some(registry) = &mut self.lsp else {
-            return;
-        };
-        if !registry.is_ready() || !registry.client().capabilities.hover_provider {
-            return;
-        }
-        let handle = self.editor.active();
-        let Some(path) = &handle.path else { return };
-        let uri = crate::lsp::types::path_to_uri(path);
-        let cursor = handle.buffer.cursors.primary();
-        let pos = crate::lsp::types::byte_offset_to_lsp_position(
-            handle.buffer.rope(),
-            cursor.byte_offset,
-        );
-        let anchor_line = cursor.line;
-        let anchor_col = cursor.col;
-
-        let _ = registry
-            .client_mut()
-            .request_hover(&uri, pos.line, pos.character);
-
-        // We'll set hover state when the response arrives.
-        let _ = (anchor_line, anchor_col);
-    }
-
-    fn apply_hover_response(&mut self, contents: Option<serde_json::Value>) {
-        let Some(contents) = contents else { return };
-        let text = extract_hover_text(&contents);
-        if text.is_empty() {
-            return;
-        }
-        let cursor = self.editor.active().buffer.cursors.primary();
-        self.hover = Some(HoverState {
-            content: text,
-            anchor_line: cursor.line,
-            anchor_col: cursor.col,
-        });
-    }
-
     // ── Go to Definition ─────────────────────────────────────────────────────
-
-    fn trigger_go_to_definition(&mut self) {
-        let Some(registry) = &mut self.lsp else {
-            return;
-        };
-        if !registry.is_ready() || !registry.client().capabilities.definition_provider {
-            return;
-        }
-        let handle = self.editor.active();
-        let Some(path) = &handle.path else { return };
-        let uri = crate::lsp::types::path_to_uri(path);
-        let cursor = handle.buffer.cursors.primary();
-        let pos = crate::lsp::types::byte_offset_to_lsp_position(
-            handle.buffer.rope(),
-            cursor.byte_offset,
-        );
-
-        let _ = registry
-            .client_mut()
-            .request_definition(&uri, pos.line, pos.character);
-    }
-
-    fn apply_definition_response(&mut self, locations: serde_json::Value) {
-        let locs = parse_locations(&locations);
-        if locs.is_empty() {
-            return;
-        }
-        if locs.len() == 1 {
-            self.jump_to_location(&locs[0]);
-        } else {
-            self.show_references_list(locs);
-        }
-    }
 
     // ── Find References ──────────────────────────────────────────────────────
 
-    fn trigger_find_references(&mut self) {
-        let Some(registry) = &mut self.lsp else {
-            return;
-        };
-        if !registry.is_ready() || !registry.client().capabilities.references_provider {
-            return;
-        }
-        let handle = self.editor.active();
-        let Some(path) = &handle.path else { return };
-        let uri = crate::lsp::types::path_to_uri(path);
-        let cursor = handle.buffer.cursors.primary();
-        let pos = crate::lsp::types::byte_offset_to_lsp_position(
-            handle.buffer.rope(),
-            cursor.byte_offset,
-        );
-
-        let _ = registry
-            .client_mut()
-            .request_references(&uri, pos.line, pos.character);
-    }
-
-    fn apply_references_response(&mut self, locations: serde_json::Value) {
-        let locs = parse_locations(&locations);
-        if locs.is_empty() {
-            return;
-        }
-        self.show_references_list(locs);
-    }
-
-    fn show_references_list(&mut self, locs: Vec<(PathBuf, usize, usize)>) {
-        let items: Vec<ReferenceItem> = locs
-            .into_iter()
-            .map(|(path, line, col)| {
-                let context = std::fs::read_to_string(&path)
-                    .ok()
-                    .and_then(|text| text.lines().nth(line).map(|l| l.trim().to_string()))
-                    .unwrap_or_default();
-                ReferenceItem {
-                    path,
-                    line,
-                    col,
-                    context,
-                }
-            })
-            .collect();
-        self.references_list = Some(ReferencesListState { items, selected: 0 });
-    }
-
-    /// Walk to the next/previous quickfix entry and jump to it. Does not
-    /// open the overlay — just navigates the list when it is already
-    /// populated. Builds a fresh list from current diagnostics if the
-    /// overlay state is empty.
-    fn quickfix_step(&mut self, step: i32) {
-        if self.quickfix.is_none() {
-            let entries = crate::quickfix::collect_lsp_diagnostics(&self.editor);
-            if entries.is_empty() {
-                self.status_error = Some("No diagnostics".into());
-                return;
-            }
-            self.quickfix = Some(crate::quickfix::QuickfixState::new(entries));
-        }
-        let entry_opt = {
-            let qf = self.quickfix.as_mut().unwrap();
-            let n = qf.entries.len();
-            if n == 0 {
-                None
-            } else {
-                let i = if step > 0 {
-                    (qf.selected + 1) % n
-                } else {
-                    (qf.selected + n - 1) % n
-                };
-                qf.selected = i;
-                Some(qf.entries[i].clone())
-            }
-        };
-        if let Some(e) = entry_opt {
-            self.jump_to_location(&(e.path, e.line, e.col));
-        }
-    }
-
-    /// Handle input while the quickfix list overlay is open.
-    fn handle_quickfix_input(&mut self, action: &EditorAction) -> bool {
-        let n = self.quickfix.as_ref().map(|q| q.entries.len()).unwrap_or(0);
-        match action {
-            EditorAction::MoveCursor(Direction::Up) => {
-                if let Some(q) = &mut self.quickfix {
-                    q.selected = q.selected.saturating_sub(1);
-                }
-                true
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                if let Some(q) = &mut self.quickfix {
-                    q.selected = (q.selected + 1).min(n.saturating_sub(1));
-                }
-                true
-            }
-            EditorAction::InsertNewline => {
-                let target = self.quickfix.as_ref().and_then(|q| {
-                    q.entries
-                        .get(q.selected)
-                        .map(|e| (e.path.clone(), e.line, e.col))
-                });
-                self.quickfix = None;
-                if let Some(loc) = target {
-                    self.jump_to_location(&loc);
-                }
-                true
-            }
-            EditorAction::CloseSearch => {
-                self.quickfix = None;
-                true
-            }
-            EditorAction::Quit | EditorAction::ForceQuit => {
-                self.quickfix = None;
-                false
-            }
-            _ => false,
-        }
-    }
-
-    /// Handle input while the clipboard-ring picker is open. Returns `true`
-    /// when the action was consumed, `false` to allow global routing.
-    fn handle_clipboard_ring(&mut self, action: &EditorAction) -> bool {
-        let num_items = self
-            .clipboard_ring
-            .as_ref()
-            .map(|r| r.entries.len())
-            .unwrap_or(0);
-        match action {
-            EditorAction::MoveCursor(Direction::Up) => {
-                if let Some(r) = &mut self.clipboard_ring {
-                    r.selected = r.selected.saturating_sub(1);
-                }
-                true
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                if let Some(r) = &mut self.clipboard_ring {
-                    r.selected = (r.selected + 1).min(num_items.saturating_sub(1));
-                }
-                true
-            }
-            EditorAction::InsertNewline => {
-                if let Some(r) = &self.clipboard_ring {
-                    let idx = r.selected;
-                    if let Some(text) = self.clipboard.pick(idx) {
-                        // Reuse the buffer's insert_str — it already handles
-                        // replacing the active selection, exactly like Paste.
-                        self.editor.active_mut().buffer.insert_str(&text);
-                    }
-                }
-                self.clipboard_ring = None;
-                true
-            }
-            EditorAction::CloseSearch => {
-                self.clipboard_ring = None;
-                true
-            }
-            EditorAction::Quit | EditorAction::ForceQuit => {
-                self.clipboard_ring = None;
-                false
-            }
-            _ => false,
-        }
-    }
-
-    fn handle_references_input(&mut self, action: &EditorAction) -> bool {
-        let num_items = self
-            .references_list
-            .as_ref()
-            .map(|r| r.items.len())
-            .unwrap_or(0);
-        match action {
-            EditorAction::MoveCursor(Direction::Up) => {
-                if let Some(r) = &mut self.references_list {
-                    r.selected = r.selected.saturating_sub(1);
-                }
-                true
-            }
-            EditorAction::MoveCursor(Direction::Down) => {
-                if let Some(r) = &mut self.references_list {
-                    r.selected = (r.selected + 1).min(num_items.saturating_sub(1));
-                }
-                true
-            }
-            EditorAction::InsertNewline => {
-                if let Some(r) = &self.references_list
-                    && let Some(item) = r.items.get(r.selected)
-                {
-                    let path = item.path.clone();
-                    let line = item.line;
-                    let col = item.col;
-                    self.references_list = None;
-                    self.jump_to_location(&(path, line, col));
-                }
-                true
-            }
-            EditorAction::CloseSearch => {
-                self.references_list = None;
-                true
-            }
-            EditorAction::Quit | EditorAction::ForceQuit => {
-                self.references_list = None;
-                false
-            }
-            _ => false,
-        }
-    }
-
-    fn jump_to_location(&mut self, loc: &(PathBuf, usize, usize)) {
-        let (path, line, col) = loc;
-
-        // Check if file is already open in a tab.
-        let existing = self
-            .editor
-            .tabs
-            .iter()
-            .position(|t| t.path.as_ref().is_some_and(|p| same_file(p, path)));
-
-        if let Some(idx) = existing {
-            self.editor.go_to_tab(idx);
-        } else {
-            // Open in new tab.
-            if self.editor.open_tab(path.clone()).is_err() {
-                return;
-            }
-            self.after_file_open_or_save();
-        }
-
-        // Jump cursor to position.
-        let rope = self.editor.active().buffer.rope().clone();
-        let cursor = crate::buffer::cursor::Cursor::from_line_col(&rope, *line, *col);
-        *self.editor.active_mut().buffer.cursors.primary_mut() = cursor;
-    }
-
     // ── Rename ───────────────────────────────────────────────────────────────
-
-    fn trigger_rename(&mut self) {
-        let Some(registry) = &self.lsp else { return };
-        if !registry.is_ready() || !registry.client().capabilities.rename_provider {
-            return;
-        }
-        // Enter rename modal: prompt for new name.
-        let handle = self.editor.active();
-        let cursor = handle.buffer.cursors.primary();
-        // Extract word under cursor as the default name.
-        let rope = handle.buffer.rope();
-        let byte = cursor.byte_offset;
-        let text = rope.to_string();
-        let word = extract_word_at(&text, byte);
-        self.input_mode = InputMode::Rename(word);
-    }
-
-    /// Send rename request after the user confirms the new name.
-    fn send_rename(&mut self, new_name: &str) {
-        let Some(registry) = &mut self.lsp else {
-            return;
-        };
-        if !registry.is_ready() {
-            return;
-        }
-        let handle = self.editor.active();
-        let Some(path) = &handle.path else { return };
-        let uri = crate::lsp::types::path_to_uri(path);
-        let cursor = handle.buffer.cursors.primary();
-        let pos = crate::lsp::types::byte_offset_to_lsp_position(
-            handle.buffer.rope(),
-            cursor.byte_offset,
-        );
-
-        let _ = registry
-            .client_mut()
-            .request_rename(&uri, pos.line, pos.character, new_name);
-    }
-
-    fn apply_workspace_edit(&mut self, edit: &serde_json::Value) {
-        let changes = match edit.get("changes").and_then(|v| v.as_object()) {
-            Some(c) => c,
-            None => return,
-        };
-
-        for (uri, edits) in changes {
-            let path = match crate::lsp::types::uri_to_path(uri) {
-                Some(p) => p,
-                None => continue,
-            };
-            let edits = match edits.as_array() {
-                Some(e) => e,
-                None => continue,
-            };
-
-            // Find or open the tab.
-            let tab_idx = self
-                .editor
-                .tabs
-                .iter()
-                .position(|t| t.path.as_ref().is_some_and(|p| same_file(p, &path)));
-            let tab_idx = match tab_idx {
-                Some(i) => i,
-                None => continue, // Skip files not open.
-            };
-
-            // Collect and sort edits in reverse order to avoid offset shifting.
-            let mut text_edits: Vec<(usize, usize, String)> = Vec::new();
-            let rope = self.editor.tabs[tab_idx].buffer.rope();
-            for e in edits {
-                let range = match e.get("range") {
-                    Some(r) => r,
-                    None => continue,
-                };
-                let start = match parse_lsp_position(range.get("start")) {
-                    Some(pos) => {
-                        crate::lsp::types::lsp_position_to_byte_offset(rope, pos).unwrap_or(0)
-                    }
-                    None => continue,
-                };
-                let end = match parse_lsp_position(range.get("end")) {
-                    Some(pos) => {
-                        crate::lsp::types::lsp_position_to_byte_offset(rope, pos).unwrap_or(0)
-                    }
-                    None => continue,
-                };
-                let new_text = e
-                    .get("newText")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
-                text_edits.push((start, end, new_text));
-            }
-
-            // Apply in reverse byte order.
-            text_edits.sort_by_key(|b| std::cmp::Reverse(b.0));
-            let tab = &mut self.editor.tabs[tab_idx];
-            for (start, end, new_text) in &text_edits {
-                let rope = tab.buffer.rope();
-                let start_char = rope.byte_to_char(*start);
-                let end_char = rope.byte_to_char(*end);
-                tab.buffer.delete_range(start_char, end_char);
-                tab.buffer.insert_str(new_text);
-            }
-        }
-    }
 
     // ── Code Action ──────────────────────────────────────────────────────────
 
-    fn trigger_code_action(&mut self) {
-        let Some(registry) = &mut self.lsp else {
-            return;
-        };
-        if !registry.is_ready() || !registry.client().capabilities.code_action_provider {
-            return;
-        }
-        let handle = self.editor.active();
-        let Some(path) = &handle.path else { return };
-        let uri = crate::lsp::types::path_to_uri(path);
-        let cursor = handle.buffer.cursors.primary();
-        let pos = crate::lsp::types::byte_offset_to_lsp_position(
-            handle.buffer.rope(),
-            cursor.byte_offset,
-        );
-        let range = serde_json::json!({
-            "start": { "line": pos.line, "character": pos.character },
-            "end": { "line": pos.line, "character": pos.character },
-        });
-
-        let _ = registry.client_mut().request_code_action(&uri, range);
-    }
-
     // ── Semantic Tokens ──────────────────────────────────────────────────────
-
-    fn apply_semantic_tokens(&mut self, uri: &str, data: &[u32]) {
-        let path = match crate::lsp::types::uri_to_path(uri) {
-            Some(p) => p,
-            None => return,
-        };
-        let tab = self
-            .editor
-            .tabs
-            .iter_mut()
-            .find(|t| t.path.as_ref().is_some_and(|p| same_file(p, &path)));
-        let Some(tab) = tab else { return };
-
-        let rope = tab.buffer.rope();
-        let tokens = crate::lsp::types::decode_semantic_tokens(data, rope);
-        tab.lsp_state.semantic_tokens = Some(tokens);
-    }
-
-    /// Request semantic tokens for the active buffer.
-    fn request_semantic_tokens_for_active(&mut self) {
-        let Some(registry) = &mut self.lsp else {
-            return;
-        };
-        if !registry.is_ready() || !registry.client().capabilities.semantic_tokens_provider {
-            return;
-        }
-        let handle = self.editor.active();
-        let Some(path) = &handle.path else { return };
-        let uri = crate::lsp::types::path_to_uri(path);
-        let _ = registry.client_mut().send_request(
-            "textDocument/semanticTokens/full",
-            Some(serde_json::json!({
-                "textDocument": { "uri": uri }
-            })),
-        );
-    }
 
     // ── LSP restart/stop ──────────────────────────────────────────────────────
 
-    fn lsp_restart(&mut self) {
-        // Tear down existing connection.
-        self.lsp = None;
-        self.pending_lsp_approval = None;
-        // Clear stale state from all buffers.
-        for tab in &mut self.editor.tabs {
-            tab.lsp_state.diagnostics.clear();
-            tab.lsp_state.semantic_tokens = None;
-        }
-        // Start fresh if config is active — routed through the trust gate.
-        self.request_lsp_start();
-    }
-
-    /// Trust-gated entry point for spawning the LSP server.
-    ///
-    /// Resolves the configured binary, hashes it, and consults the user-global
-    /// trust store. If the binary is approved, spawns directly. If unknown or
-    /// the hash has changed, sets `pending_lsp_approval` so the approval
-    /// overlay opens on the next frame.
-    pub fn request_lsp_start(&mut self) {
-        if std::env::var_os("TXT_DISABLE_LSP").is_some() {
-            return;
-        }
-        if !self.lsp_config.is_active() {
-            return;
-        }
-        let entry = match self.lsp_config.active_server() {
-            Some(e) => e.clone(),
-            None => return,
-        };
-        let server_name = self.lsp_config.server.clone().unwrap_or_default();
-
-        let resolved = match crate::lsp::resolve::resolve_binary(&entry.command) {
-            Ok(r) => r,
-            Err(e) => {
-                self.status_error = Some(format!("LSP: {e}"));
-                return;
-            }
-        };
-        let hash = match crate::lsp::resolve::hash_file(&resolved.canonical_path) {
-            Ok(h) => h,
-            Err(e) => {
-                self.status_error = Some(format!("LSP: {e}"));
-                return;
-            }
-        };
-
-        let store = crate::lsp::trust::TrustStore::load();
-        match store.check(&resolved.canonical_path, &hash) {
-            crate::lsp::trust::TrustDecision::Approved => {
-                self.lsp = crate::lsp::LspRegistry::start(&self.lsp_config, &self.workspace).ok();
-            }
-            crate::lsp::trust::TrustDecision::Unknown => {
-                self.pending_lsp_approval = Some(PendingLspApproval {
-                    server_name,
-                    command: entry.command.clone(),
-                    args: entry.args.clone(),
-                    display_path: resolved.display_path,
-                    canonical_path: resolved.canonical_path,
-                    hash,
-                    reason: LspApprovalReason::FirstLaunch,
-                });
-            }
-            crate::lsp::trust::TrustDecision::HashMismatch { previous_hash } => {
-                self.pending_lsp_approval = Some(PendingLspApproval {
-                    server_name,
-                    command: entry.command.clone(),
-                    args: entry.args.clone(),
-                    display_path: resolved.display_path,
-                    canonical_path: resolved.canonical_path,
-                    hash,
-                    reason: LspApprovalReason::BinaryChanged { previous_hash },
-                });
-            }
-        }
-    }
-
-    /// Whether the currently configured LSP binary still matches its trust-store
-    /// entry. Used by the crash-recovery path to decide between an in-place
-    /// restart and re-routing through the approval gate.
-    fn lsp_binary_still_trusted(&self) -> bool {
-        let entry = match self.lsp_config.active_server() {
-            Some(e) => e,
-            None => return false,
-        };
-        let resolved = match crate::lsp::resolve::resolve_binary(&entry.command) {
-            Ok(r) => r,
-            Err(_) => return false,
-        };
-        let hash = match crate::lsp::resolve::hash_file(&resolved.canonical_path) {
-            Ok(h) => h,
-            Err(_) => return false,
-        };
-        matches!(
-            crate::lsp::trust::TrustStore::load().check(&resolved.canonical_path, &hash),
-            crate::lsp::trust::TrustDecision::Approved
-        )
-    }
-
-    /// Handle input while the LSP-binary approval overlay is shown.
-    /// Returns `true` if the action was consumed (always, while modal is open).
-    fn handle_lsp_approval(&mut self, action: &EditorAction) -> bool {
-        match action {
-            EditorAction::InsertChar('y') | EditorAction::InsertChar('Y') => {
-                if let Some(pending) = self.pending_lsp_approval.take() {
-                    let mut store = crate::lsp::trust::TrustStore::load();
-                    store.approve(
-                        pending.canonical_path,
-                        pending.hash,
-                        Some(pending.server_name),
-                    );
-                    store.save();
-                    self.lsp =
-                        crate::lsp::LspRegistry::start(&self.lsp_config, &self.workspace).ok();
-                    self.status_error = Some(
-                        "LSP approved; recorded in ~/.config/txt/trusted_binaries.json".into(),
-                    );
-                }
-                true
-            }
-            EditorAction::InsertChar('n')
-            | EditorAction::InsertChar('N')
-            | EditorAction::Quit
-            | EditorAction::CloseSearch => {
-                self.pending_lsp_approval = None;
-                self.status_error = Some("LSP not started.".into());
-                true
-            }
-            // Capture every other input while the modal is up.
-            _ => true,
-        }
-    }
-
     // ── Coordinate helpers ───────────────────────────────────────────────────
-
-    /// Try to handle `c` as a bracket-pair insertion or skip-on-close.
-    /// Returns `true` when this method consumed the keystroke; the caller
-    /// must skip the normal insert path. Caller already ensured the buffer
-    /// is single-cursor and `config.auto_pair` is on.
-    fn try_auto_pair(&mut self, c: char) -> bool {
-        // Pair table — opener → closer. Symmetric pairs (`"` etc.) close
-        // with the same character.
-        let pair_for = |ch: char| -> Option<char> {
-            match ch {
-                '(' => Some(')'),
-                '[' => Some(']'),
-                '{' => Some('}'),
-                '"' => Some('"'),
-                '\'' => Some('\''),
-                '`' => Some('`'),
-                _ => None,
-            }
-        };
-        let close_for = |ch: char| -> bool { matches!(ch, ')' | ']' | '}' | '"' | '\'' | '`') };
-
-        let buf = &mut self.editor.active_mut().buffer;
-        let cursor = buf.cursors.primary();
-        if cursor.has_selection() {
-            return false;
-        }
-        let at = cursor.byte_offset;
-        let rope = buf.rope();
-        let next_char = if at < rope.len_bytes() {
-            let ch_idx = rope.byte_to_char(at);
-            rope.chars_at(ch_idx).next()
-        } else {
-            None
-        };
-        let prev_char = if at > 0 {
-            let ch_idx = rope.byte_to_char(at);
-            // Walk one char back.
-            let mut iter = rope.chars_at(ch_idx);
-            iter.prev()
-        } else {
-            None
-        };
-
-        // Skip-on-close: typing `)` over an existing `)` just advances.
-        if close_for(c) && next_char == Some(c) {
-            let new_off = at + c.len_utf8();
-            buf.move_cursor_to(new_off, false);
-            return true;
-        }
-
-        // Auto-open: insert `(` `)` and step back. Symmetric pairs need extra
-        // care to avoid pairing a closing quote with itself when the cursor
-        // is right after a word character (e.g. `let's` → don't pair the
-        // apostrophe).
-        if let Some(close) = pair_for(c) {
-            let is_symmetric = c == close;
-            if is_symmetric {
-                // Don't auto-pair quotes when adjacent to a word char on
-                // either side (prevents the `let's` case and contractions).
-                let is_word = |ch: Option<char>| match ch {
-                    Some(c) => c.is_alphanumeric() || c == '_',
-                    None => false,
-                };
-                if is_word(prev_char) || is_word(next_char) {
-                    return false;
-                }
-            }
-            let mut pair = String::with_capacity(c.len_utf8() + close.len_utf8());
-            pair.push(c);
-            pair.push(close);
-            buf.insert_str(&pair);
-            // Step the cursor back one char so the user types inside the pair.
-            let after = buf.cursors.primary().byte_offset;
-            let target = after.saturating_sub(close.len_utf8());
-            buf.move_cursor_to(target, false);
-            return true;
-        }
-        false
-    }
-
-    /// Wrap the active selection (or word at the cursor) in the delimiter
-    /// pair chosen by `ch`. Symmetric punctuation (e.g. `"`, `'`) uses the
-    /// same character on both sides.
-    fn apply_surround(&mut self, ch: char) {
-        let pair: Option<(String, String)> = match ch {
-            '(' | ')' => Some(("(".into(), ")".into())),
-            '[' | ']' => Some(("[".into(), "]".into())),
-            '{' | '}' => Some(("{".into(), "}".into())),
-            '<' | '>' => Some(("<".into(), ">".into())),
-            '"' | '\'' | '`' => Some((ch.to_string(), ch.to_string())),
-            c if c.is_ascii_punctuation() => Some((c.to_string(), c.to_string())),
-            _ => None,
-        };
-        let Some((open, close)) = pair else {
-            self.status_error = Some(format!("No surround pair for '{ch}'"));
-            return;
-        };
-        let ok = self
-            .editor
-            .active_mut()
-            .buffer
-            .surround_selection(&open, &close);
-        if !ok {
-            self.status_error = Some("No selection or word to surround".into());
-        }
-    }
-
-    fn selected_text(&self) -> Option<String> {
-        let cursor = self.editor.active().buffer.cursors.primary();
-        if !cursor.has_selection() {
-            return None;
-        }
-        let range = cursor.selection_bytes();
-        let start = self.editor.active().buffer.rope().byte_to_char(range.start);
-        let end = self.editor.active().buffer.rope().byte_to_char(range.end);
-        Some(
-            self.editor
-                .active()
-                .buffer
-                .rope()
-                .slice(start..end)
-                .to_string(),
-        )
-    }
-
-    /// Hit-test the tab strip. Returns the tab index when `(col, row)`
-    /// lands on a rendered tab label.
-    fn tab_bar_tab_at(&self, col: u16, row: u16) -> Option<usize> {
-        let area = self.tab_bar_area?;
-        crate::ui::tab_bar::tab_at(&self.editor, area, col, row)
-    }
-
-    /// Cap user-initiated scrolling so the last line cannot pass the vertical
-    /// centre of the viewport and the longest line's end cannot pass the
-    /// horizontal centre. Mirrors `Viewport::clamp_user_scroll` but supplies
-    /// the buffer-derived dimensions from current `AppState`.
-    fn clamp_viewport_scroll(&mut self, text_h: usize) {
-        let total_lines = self.editor.active().buffer.len_lines();
-        let label = self.version_badge();
-        let gutter = effective_gutter_width(total_lines, label.as_deref());
-        let sidebar_w: u16 = if self.sidebar.is_some() {
-            self.sidebar_width + 1
-        } else {
-            0
-        };
-        let text_w =
-            (self.term_width as usize).saturating_sub(gutter as usize + 1 + sidebar_w as usize);
-
-        let tab = self.editor.active_mut();
-        let longest = crate::editor::viewport::max_line_display_width(&tab.buffer);
-        tab.viewport
-            .clamp_user_scroll(total_lines, longest, text_h, text_w);
-    }
-
-    /// Returns `true` when `action` is a vertical scroll that the viewport
-    /// (or sidebar) can't honour because it's already at the corresponding
-    /// edge. Used by the run loop to drop end-of-buffer wheel spam before
-    /// it ever reaches `update` and triggers a redraw.
-    ///
-    /// Horizontal and non-scroll actions always return `false` — the
-    /// horizontal case would need an O(n) max-line-width scan per event,
-    /// which is more expensive than letting `update` handle it normally.
-    pub(crate) fn scroll_action_is_no_op(
-        &self,
-        action: &EditorAction,
-        terminal_height: u16,
-    ) -> bool {
-        let dir = match action {
-            EditorAction::Scroll(d) => *d,
-            EditorAction::MouseScroll { dir, col, row } => {
-                if self.point_in_sidebar(*col, *row) {
-                    return self.sidebar_scroll_is_no_op(*dir);
-                }
-                *dir
-            }
-            _ => return false,
-        };
-
-        // Match the value `update` derives for scroll handlers.
-        let text_h = (terminal_height as usize).saturating_sub(1);
-        let tab = self.editor.active();
-        let vp = &tab.viewport;
-        let total_lines = tab.buffer.len_lines();
-        let max_row = total_lines.saturating_sub(1).saturating_sub(text_h / 2);
-
-        match dir {
-            ScrollDir::Up | ScrollDir::HalfPageUp => vp.scroll_row == 0,
-            ScrollDir::Down | ScrollDir::HalfPageDown => vp.scroll_row >= max_row,
-            ScrollDir::Left | ScrollDir::Right => false,
-        }
-    }
-
-    /// Whether a wheel-scroll over the sidebar would be a no-op (already at
-    /// top or already past the end of the entry list).
-    fn sidebar_scroll_is_no_op(&self, dir: ScrollDir) -> bool {
-        let sb = match self.sidebar.as_ref() {
-            Some(sb) => sb,
-            None => return true,
-        };
-        let h = self.sidebar_area.map(|r| r.height as usize).unwrap_or(0);
-        let max = if h == 0 || sb.entries.len() <= h {
-            0
-        } else {
-            sb.entries.len() - h
-        };
-        match dir {
-            ScrollDir::Up => sb.scroll_offset == 0,
-            ScrollDir::Down => sb.scroll_offset >= max,
-            _ => true,
-        }
-    }
-
-    /// Returns true if the given screen point is inside the sidebar's
-    /// entry-list area (excludes the separator column).
-    fn point_in_sidebar(&self, col: u16, row: u16) -> bool {
-        match self.sidebar_area {
-            Some(area) => {
-                col >= area.x
-                    && col < area.x + self.sidebar_width
-                    && row >= area.y
-                    && row < area.y + area.height
-            }
-            None => false,
-        }
-    }
-
-    /// Returns true if the given screen point is on the sidebar's separator
-    /// column (the 1-column-wide vertical bar between sidebar and editor).
-    fn point_on_separator(&self, col: u16, row: u16) -> bool {
-        match self.sidebar_area {
-            Some(area) => {
-                col == area.x + self.sidebar_width && row >= area.y && row < area.y + area.height
-            }
-            None => false,
-        }
-    }
-
-    /// Map a screen `row` inside the sidebar to the corresponding entry index.
-    /// Returns `None` if the row is outside the sidebar or past the last entry.
-    fn sidebar_entry_at(&self, row: u16) -> Option<usize> {
-        let area = self.sidebar_area?;
-        if row < area.y || row >= area.y + area.height {
-            return None;
-        }
-        let sb = self.sidebar.as_ref()?;
-        let screen_row = (row - area.y) as usize;
-        let idx = sb.scroll_offset + screen_row;
-        if idx < sb.entries.len() {
-            Some(idx)
-        } else {
-            None
-        }
-    }
-
-    /// Sync the sidebar's `scroll_offset` so the selected entry remains visible.
-    /// Called after any keyboard navigation that may move `selected` off-screen.
-    fn ensure_sidebar_selected_visible(&mut self) {
-        let h = self.sidebar_area.map(|r| r.height as usize).unwrap_or(0);
-        if let Some(sb) = &mut self.sidebar {
-            sb.ensure_selected_visible(h);
-        }
-    }
-
-    fn screen_to_byte(&self, col: u16, row: u16) -> Option<usize> {
-        let (adjusted_col, editor_area_y, gutter_cols, text_width) =
-            self.screen_mouse_geometry(col)?;
-        Some(screen_pos_to_byte_offset(
-            adjusted_col,
-            row,
-            editor_area_y,
-            gutter_cols,
-            text_width,
-            &self.editor.active().buffer,
-            &self.editor.active().viewport,
-        ))
-    }
-
-    /// Convert a screen position into `(line, display_col)` for box selection.
-    /// Returns `None` if the click landed in the sidebar.
-    fn screen_to_line_col(&self, col: u16, row: u16) -> Option<(usize, usize)> {
-        let (adjusted_col, editor_area_y, gutter_cols, text_width) =
-            self.screen_mouse_geometry(col)?;
-        Some(screen_pos_to_line_display_col(
-            adjusted_col,
-            row,
-            editor_area_y,
-            gutter_cols,
-            text_width,
-            &self.editor.active().buffer,
-            &self.editor.active().viewport,
-        ))
-    }
-
-    /// Shared geometry for mouse-to-buffer conversion: returns
-    /// `(adjusted_col, editor_area_y, gutter_cols, text_width)`, or `None`
-    /// when the click is inside the sidebar. `text_width` matches the column
-    /// width the renderer uses for word wrap.
-    fn screen_mouse_geometry(&self, col: u16) -> Option<(u16, u16, u16, u16)> {
-        let editor_area_y: u16 = if self.editor.tab_count() > 1 { 1 } else { 0 };
-        let sidebar_offset: u16 = if self.sidebar.is_some() {
-            self.sidebar_width + 1
-        } else {
-            0
-        };
-        if self.sidebar.is_some() && col < sidebar_offset {
-            return None;
-        }
-        let adjusted_col = col.saturating_sub(sidebar_offset);
-        let handle = self.editor.active();
-        let total_lines = handle.buffer.len_lines();
-        let label = self.version_badge();
-        let gutter = effective_gutter_width(total_lines, label.as_deref());
-        let git_col_w: u16 = if self.git_gutter.is_some() {
-            crate::ui::editor_view::GIT_GUTTER_W
-        } else {
-            0
-        };
-        let diag_col_w: u16 = if !handle.lsp_state.diagnostics.is_empty() {
-            crate::ui::editor_view::DIAG_GUTTER_W
-        } else {
-            0
-        };
-        let has_folds = (0..total_lines).any(|i| handle.folds.is_fold_start_candidate(i));
-        let fold_col_w: u16 = if has_folds {
-            crate::ui::editor_view::FOLD_GUTTER_W
-        } else {
-            0
-        };
-        let gutter_cols =
-            git_col_w + diag_col_w + fold_col_w + gutter + crate::ui::editor_view::GUTTER_PAD;
-        let editor_area_w = self.term_width.saturating_sub(sidebar_offset);
-        let text_width = editor_area_w.saturating_sub(gutter_cols);
-        Some((adjusted_col, editor_area_y, gutter_cols, text_width))
-    }
 }
 
 // ── App (top-level runner) ────────────────────────────────────────────────────
